% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC.R
\name{AICc}
\alias{AICc}
\title{Get the AICc value or table.}
\usage{
AICc(object, ...)
}
\arguments{
\item{object}{The object.}

\item{...}{Optional.}
}
\description{
Get the AICc value or table.
}
\examples{
AICc(boron_lnorm)
AICc(boron_dists)
}
