% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{augment.fitdists}
\alias{augment.fitdists}
\title{fitdists Object's Data}
\usage{
\method{augment}{fitdists}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused.}
}
\value{
A tibble of a fitdist's object's data set.
}
\description{
Gets a tibble of fitdists object's data.
}
\examples{
fits <- ssd_fit_dists(ssdtools::boron_data)
augment(fits)
}
\seealso{
Other generics: 
\code{\link{glance.fitdists}()},
\code{\link{tidy.fitdists}()}
}
\concept{generics}
