% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{augment.tmbfit}
\alias{augment.tmbfit}
\title{Construct a single row tibble::tibble() "glance" of a tmbfit model.}
\usage{
\method{augment}{tmbfit}(x, ...)
}
\arguments{
\item{x}{A tmbfit object to be converted into a single row tibble::tibble.}

\item{...}{Unused.}
}
\description{
Construct a single row tibble::tibble() "glance" of a tmbfit model.
}
