% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.fitdists}
\alias{autoplot.fitdists}
\title{Plot a fitdists Object}
\usage{
\method{autoplot}{fitdists}(object, ...)
}
\arguments{
\item{object}{The object.}

\item{...}{Unused.}
}
\value{
A ggplot object.
}
\description{
A wrapper on \code{\link[=ssd_plot_cdf]{ssd_plot_cdf()}}.
}
\examples{
fits <- ssd_fit_dists(ssddata::ccme_boron)
autoplot(fits)
}
\seealso{
\code{\link[=ssd_plot_cdf]{ssd_plot_cdf()}}
}
