% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.fitdistscens}
\alias{autoplot.fitdistscens}
\title{Autoplot}
\usage{
\method{autoplot}{fitdistscens}(
  object,
  xlab = "Concentration",
  ylab = "Species Affected",
  ic = "aic",
  ...
)
}
\arguments{
\item{object}{The object to plot.}

\item{xlab}{A string of the x-axis label.}

\item{ylab}{A string of the x-axis label.}

\item{ic}{A string indicating which information-theoretic criterion ('aic', 'aicc' or 'bic') to use for model averaging.}

\item{...}{Unused.}
}
\description{
Autoplot
}
\examples{
\dontrun{
ggplot2::autoplot(boron_dists)
}
}
