% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot}
\alias{autoplot}
\alias{autoplot.fitdist}
\alias{autoplot.fitdists}
\alias{autoplot.fitdistcens}
\title{Autoplot}
\usage{
autoplot(object, ...)

\method{autoplot}{fitdist}(object, ...)

\method{autoplot}{fitdists}(object, ...)

\method{autoplot}{fitdistcens}(object, ...)
}
\arguments{
\item{object}{The object.}

\item{...}{Unused.}
}
\description{
Plots the cumulative distribution function (cdf) using the ggplot2
generic.
}
\section{Methods (by class)}{
\itemize{
\item \code{fitdist}: Autoplot fitdist

\item \code{fitdists}: Autoplot fitdists

\item \code{fitdistcens}: Autoplot fitdistcens
}}

\examples{
ggplot2::autoplot(boron_lnorm)
ggplot2::autoplot(boron_dists)
fluazinam_lnorm$censdata$right[3] <- fluazinam_lnorm$censdata$left[3] * 1.5
fluazinam_lnorm$censdata$left[5] <- NA
ggplot2::autoplot(fluazinam_lnorm)
}
\seealso{
\code{\link[=ssd_plot_cdf]{ssd_plot_cdf()}}
}
