% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boron_all}
\alias{boron_all}
\title{fitdists for CCME Boron Data}
\format{An object of class \code{fitdists} of length 8.}
\usage{
boron_all
}
\description{
A fitdists object for Species Sensitivity Data for
Boron.
}
\examples{
boron_all
}
\keyword{datasets}
