% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boron_data}
\alias{boron_data}
\title{CCME Species Sensitivity Data for Boron}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 28 rows and 5 columns.
}
\usage{
boron_data
}
\description{
Species Sensitivity Data from
the Canadian Council of Ministers of the Environment.
}
\details{
Additional information is available from \url{http://ceqg-rcqe.ccme.ca/download/en/324/}.

The columns are as follows

\describe{
\item{Chemical}{The chemical (chr).}
\item{Species}{The species binomial name (chr).}
\item{Concentration}{The chemical concentration (dbl).}
\item{Units}{The units (chr).}
\item{Group}{The taxonomic group (fctr).}
}
}
\examples{
head(ccme_data)
}
\seealso{
\link{ccme_data}

Other boron: 
\code{\link{boron_dists}},
\code{\link{boron_hc5}},
\code{\link{boron_pred}}
}
\concept{boron}
\keyword{datasets}
