% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boron_dists}
\alias{boron_dists}
\title{fitdists for CCME Boron Data}
\format{An object of class \code{fitdists} of length 3.}
\usage{
boron_dists
}
\description{
A fitdists object for Species Sensitivity Data for
Boron.
}
\examples{
boron_dists
}
\keyword{datasets}
