% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boron_lnorm}
\alias{boron_lnorm}
\title{fitdist for CCME Boron Data}
\format{
An object of class \code{fitdist} of length 17.
}
\usage{
boron_lnorm
}
\description{
A fitdist object for Species Sensitivity Data for
Boron with the lnorm distribution.
}
\examples{
boron_lnorm
}
\seealso{
Other boron: 
\code{\link{boron_data}},
\code{\link{boron_dists}},
\code{\link{boron_hc5}},
\code{\link{boron_pred}}
}
\concept{boron}
\keyword{datasets}
