% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burr.R
\name{burr}
\alias{burr}
\alias{dburr}
\alias{qburr}
\alias{pburr}
\alias{rburr}
\alias{mburr}
\title{Burr Distribution}
\usage{
dburr(x, shape1, shape2, rate = 1, scale = 1/rate, log = FALSE)

qburr(p, shape1, shape2, rate = 1, scale = 1/rate, lower.tail = TRUE,
  log.p = FALSE)

pburr(q, shape1, shape2, rate = 1, scale = 1/rate, lower.tail = TRUE,
  log.p = FALSE)

rburr(n, shape1, shape2, rate = 1, scale = 1/rate)

mburr(order, shape1, shape2, rate = 1, scale = 1/rate)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{shape1, shape2, scale}{parameter. Must be strictly positive.}

\item{rate}{an alternative way to specify the scale.}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{p}{vector of probabilities.}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X <= x],otherwise, P[X > x].}

\item{n}{number of observations.}

\item{order}{order of the moment.}
}
\value{
dburr gives the density, pburr gives the distribution function,
qburr gives the quantile function, rburr generates random deviates and mburr
gives the kth raw moment.
}
\description{
Density, distribution function, quantile function and random generation
for the Burr distribution with \code{shape1}, \code{shape2} and \code{scale} parameters.
}
\details{
The functions are wrappers to export the identical functions from the actuar package.
}
\examples{
x <- rburr(1000,10,10)
hist(x,freq=FALSE,col='gray',border='white')
curve(dburr(x,10,10),add=TRUE,col='red4',lwd=2)
}
\seealso{
\code{\link[actuar]{burr}}
}
