% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burrIII2.R
\name{burrIII2}
\alias{burrIII2}
\alias{dburrIII2}
\alias{qburrIII2}
\alias{pburrIII2}
\alias{rburrIII2}
\alias{sburrIII2}
\title{Burr Type III Two-Parameter Distribution}
\usage{
dburrIII2(x, lshape1 = 0, lshape2 = 0, log = FALSE)

qburrIII2(p, lshape1 = 0, lshape2 = 0, lower.tail = TRUE, log.p = FALSE)

pburrIII2(q, lshape1 = 0, lshape2 = 0, lower.tail = TRUE, log.p = FALSE)

rburrIII2(n, lshape1 = 0, lshape2 = 0)

sburrIII2(x)
}
\arguments{
\item{x}{A numeric vector of values.}

\item{lshape1}{shape1 parameter on the log scale.}

\item{lshape2}{shape2 parameter on the log scale.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}

\item{p}{vector of probabilities.}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X <= x],otherwise, P[X > x].}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{q}{vector of quantiles.}

\item{n}{number of observations.}
}
\value{
A numeric vector.
}
\description{
Density, distribution function, quantile function, random generation
and starting values for the
Burr Type III two-parameter distribution
with \code{lshape} and \code{lscale} parameters.
}
\examples{
x <- seq(0.01, 5, by = 0.01)
plot(x, dburrIII2(x), type = "l")
}
\seealso{
\code{\link{burrIII3}}
}
