% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burrIII3b.R
\name{burrIII3b}
\alias{burrIII3b}
\alias{dburrIII3b}
\alias{pburrIII3b}
\alias{qburrIII3b}
\alias{rburrIII3b}
\alias{sburrIII3b}
\title{Burr Type III Three-Parameter Distribution}
\usage{
dburrIII3b(x, lshape1 = 0, lshape2 = 0, lscale = 0, log = FALSE)

pburrIII3b(
  q,
  lshape1 = 0,
  lshape2 = 0,
  lscale = 0,
  lower.tail = TRUE,
  log.p = FALSE
)

qburrIII3b(
  p,
  lshape1 = 0,
  lshape2 = 0,
  lscale = 0,
  lower.tail = TRUE,
  log.p = FALSE
)

rburrIII3b(n, lshape1 = 0, lshape2 = 0, lscale = 0)

sburrIII3b(x)
}
\arguments{
\item{x}{A numeric vector of values.}

\item{lshape1}{shape1 parameter on the log scale.}

\item{lshape2}{shape2 parameter on the log scale.}

\item{lscale}{scale parameter on the log scale.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}

\item{q}{vector of quantiles.}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X <= x],otherwise, P[X > x].}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
A numeric vector.
}
\description{
Probability density, cumulative distribution, 
inverse cumulative distribution, random sample and starting values functions.
}
\details{
The Burr III distribution is the distribution of 1/x where 
x has the Burr Type 12 distribution 
refer to 
https://www.itl.nist.gov/div898/software/dataplot/refman2/auxillar/bu3pdf.htm 
for details.
}
