% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burrrIII3o.R
\name{burrIII3o}
\alias{burrIII3o}
\alias{dburrIII3o}
\alias{qburrIII3o}
\alias{pburrIII3o}
\alias{rburrIII3o}
\title{Burr Type III Three-Parameter Distribution}
\usage{
dburrIII3o(x, lshape1 = log(1), lshape2 = log(1), lscale = log(1), log = FALSE)

qburrIII3o(
  p,
  lshape1 = log(1),
  lshape2 = log(1),
  lscale = log(1),
  lower.tail = TRUE,
  log.p = FALSE
)

pburrIII3o(
  q,
  lshape1 = log(1),
  lshape2 = log(1),
  lscale = log(1),
  lower.tail = TRUE,
  log.p = FALSE
)

rburrIII3o(n, lshape1 = log(1), lshape2 = log(1), lscale = log(1))
}
\arguments{
\item{x}{The object.}

\item{lshape1}{shape1 parameter on the log scale.}

\item{lshape2}{shape2 parameter on the log scale.}

\item{lscale}{scale parameter on the log scale.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}

\item{p}{vector of probabilities.}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X <= x],otherwise, P[X > x].}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{q}{vector of quantiles.}

\item{n}{number of observations.}
}
\value{
dburrIII3o gives the density, pburrIII3o gives the distribution function,
qburrIII3o gives the quantile function, and rburrIII3o generates random samples.
}
\description{
Density, distribution function, quantile function and random generation
for the Burr Type III Three-Parameter distribution with 
\code{lshape} and \code{lscale} parameters.
}
\details{
The Burr 12 distribution from the actuar package is used as a base.
The Burr III distribution is the distribution of 1/x where x has the Burr Type 12 distribution.
refer to https://www.itl.nist.gov/div898/software/dataplot/refman2/auxillar/bu3pdf.htm for details.
The shape1, shape2, and scale paramters are on the log(scale) as these must be positive.
}
\examples{
x <- rburrIII3o(1000)
hist(x, freq = FALSE, col = "gray", border = "white")
curve(dburrIII3o(x), add = TRUE, col = "red4", lwd = 2)
}
\seealso{
\code{\link[actuar]{dburr}}
}
