% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ccme_data}
\alias{ccme_data}
\title{CCME Species Sensitivity Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 144 rows and 5 columns.
}
\usage{
ccme_data
}
\description{
Species Sensitivity Data from
the Canadian Council of Ministers of the Environment.
The taxonomic groups are Amphibian, Fish, Invertebrate and Plant.
Plants includes freshwater algae.
}
\details{
Additional information on each of the chemicals is available from the
CCME website.
\describe{
\item{Boron}{\url{http://ceqg-rcqe.ccme.ca/download/en/324/}}
\item{Cadmium}{\url{http://ceqg-rcqe.ccme.ca/download/en/148/}}
\item{Chloride}{\url{http://ceqg-rcqe.ccme.ca/download/en/337/}}
\item{Endosulfan}{\url{http://ceqg-rcqe.ccme.ca/download/en/327/}}
\item{Glyphosate}{\url{http://ceqg-rcqe.ccme.ca/download/en/182/}}
\item{Uranium}{\url{http://ceqg-rcqe.ccme.ca/download/en/328/}}
\item{Silver}{\url{http://ceqg-rcqe.ccme.ca/download/en/355/}}
}

\describe{
\item{Chemical}{The chemical (chr).}
\item{Species}{The species binomial name (chr).}
\item{Conc}{The chemical concentration (dbl).}
\item{Group}{The taxonomic group (fctr).}
\item{Units}{The units (chr).}
}
}
\examples{
head(ccme_data)
}
\keyword{datasets}
