% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.fitdists}
\alias{coef.fitdists}
\title{Turn a fitdists Object into a Tidy Tibble}
\usage{
\method{coef}{fitdists}(object, ...)
}
\arguments{
\item{object}{The object.}

\item{...}{Unused.}
}
\description{
A wrapper on \code{\link[=tidy.fitdists]{tidy.fitdists()}}.
}
\examples{
fits <- ssd_fit_dists(ssddata::ccme_boron)
coef(fits)
}
\seealso{
\code{\link[=tidy.fitdists]{tidy.fitdists()}}
}
