% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{comma_signif}
\alias{comma_signif}
\title{Comma and Significance Formatter
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
comma_signif(x, digits = 3, ..., big.mark = ",")
}
\arguments{
\item{x}{A numeric vector to format.}

\item{digits}{A whole number specifying the number of significant figures.}

\item{...}{Unused.}

\item{big.mark}{A string specifying the thousands separator.}
}
\value{
A character vector.
}
\description{
Deprecated for \code{ssd_label_comma()}
}
\examples{
\dontrun{
comma_signif(c(0.1, 1, 10, 1000, 10000))
}
}
\seealso{
\code{\link[=ssd_label_comma]{ssd_label_comma()}}
}
\keyword{internal}
