% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgumbel.R
\name{dlgumbel}
\alias{dlgumbel}
\title{Log-Gumbel (Inverse Weibull) Probability Density}
\usage{
dlgumbel(x, locationlog = 0, scalelog = 1, log = FALSE)
}
\arguments{
\item{x}{A numeric vector of values.}

\item{locationlog}{location on the log scale parameter.}

\item{scalelog}{scale on log scale parameter.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
A numeric vector.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
