% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{estimates.fitdists}
\alias{estimates.fitdists}
\title{Estimates for fitdists Object}
\usage{
\method{estimates}{fitdists}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused.}
}
\value{
A named list of the estimates.
}
\description{
Gets a named list of the estimate values by term and distribution.
}
\examples{
fits <- ssd_fit_dists(ssdtools::boron_data)
estimates(fits)
}
\seealso{
\code{\link[=tidy.fitdists]{tidy.fitdists()}}
}
