% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_dist_internal}
\alias{fit_dist_internal}
\title{Fit Distribution}
\usage{
fit_dist_internal(x, dist = "lnorm")
}
\arguments{
\item{x}{A numeric vector of the data.}

\item{dist}{A string of the distribution to fit.}
}
\value{
An object of class fitdist.
}
\description{
Fit Distribution
}
