% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fluazinam_lnorm}
\alias{fluazinam_lnorm}
\title{fitdist for CCME Fluazinam Data}
\format{
An object of class \code{fitdistcens} of length 17.
}
\usage{
fluazinam_lnorm
}
\description{
A fitdist object for Species Sensitivity Data for
Boron with the lnorm distribution.
}
\examples{
fluazinam_lnorm
}
\seealso{
\code{\link[fitdistrplus:fluazinam]{fitdistrplus::fluazinam()}}

Other fluazinam: 
\code{\link{fluazinam_dists}},
\code{\link{fluazinam_pred}}
}
\concept{fluazinam}
\keyword{datasets}
