% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma.R
\name{gamma}
\alias{gamma}
\alias{sgamma}
\title{Gamma Distribution}
\usage{
sgamma(x)
}
\arguments{
\item{x}{A numeric vector of values.}
}
\value{
A numeric vector.
}
\description{
Starting values for the
Gamma distribution.
}
\examples{
x <- rgamma(1000, 1, 0.1)
hist(log(x), freq = FALSE, col = "gray", border = "white")
hist(x, freq = FALSE, col = "gray", border = "white")
curve(dgamma(x, 1, 0.1), add = TRUE, col = "red4", lwd = 2)
}
\seealso{
\code{\link[stats]{dgamma}}
}
