% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_ADFun.R
\name{gamma_ADFun}
\alias{gamma_ADFun}
\title{Create a \code{TMB::ADFun} object for the gamma likelihood.}
\usage{
gamma_ADFun(x)
}
\arguments{
\item{x}{Vector of observations.}
}
\value{
A list as returned by \code{TMB::MakeADFun} representing the negative loglikelihood of a gamma distribution.
}
\description{
Create a \code{TMB::ADFun} object for the gamma likelihood.
}
