% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R, R/ggplot2-ggproto.R
\docType{data}
\name{StatSsd}
\alias{StatSsd}
\alias{StatCdf}
\alias{GeomSsd}
\alias{ggplot2-ggproto}
\title{Base ggproto classes for ggplot2}
\description{
If you are creating a new geom, stat, position, or scale in another package,
you'll need to extend from `ggplot2::Geom`, `ggplot2::Stat`,
`ggplot2::Position`, or `ggplot2::Scale`.
}
\seealso{
ggproto
}
\keyword{datasets}
\keyword{internal}
