% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glance.R
\name{glance.fitdist}
\alias{glance.fitdist}
\title{Construct a single row tibble::tibble() "glance" of a fitdist model.}
\usage{
\method{glance}{fitdist}(x, ...)
}
\arguments{
\item{x}{A tmbfit object to be converted into a single row tibble::tibble.}

\item{...}{Unused.}
}
\description{
Construct a single row tibble::tibble() "glance" of a fitdist model.
}
