% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glance.R
\name{glance.fitdists}
\alias{glance.fitdists}
\title{Construct a tibble::tibble() with a single row "glance" for each tmbfit model.}
\usage{
\method{glance}{fitdists}(x, ...)
}
\arguments{
\item{x}{A fitdists object to be converted into a tidy tibble::tibble().}

\item{...}{Must be unused.}
}
\description{
Turns a fitdists object into a glance tibble.
}
