% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.fitdistcens}
\alias{is.fitdistcens}
\title{Is censored fitdist}
\usage{
is.fitdistcens(x)
}
\arguments{
\item{x}{The object.}
}
\value{
A flag.
}
\description{
Tests whether an object is a censored fitdist.
}
\examples{
is.fitdistcens(boron_lnorm)
is.fitdistcens(fluazinam_lnorm)
}
