% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.fitdistscens}
\alias{is.fitdistscens}
\title{Is censored fitdists}
\usage{
is.fitdistscens(x)
}
\arguments{
\item{x}{The object.}
}
\value{
A flag.
}
\description{
Tests whether an object is a censored fitdists.
}
\examples{
is.fitdistscens(boron_dists)
is.fitdistscens(fluazinam_lnorm)
is.fitdistscens(fluazinam_dists)
}
\seealso{
Other is: 
\code{\link{is.fitdistcens}()},
\code{\link{is.fitdists}()},
\code{\link{is.fitdist}()}
}
\concept{is}
