% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.tmbfit}
\alias{is.tmbfit}
\title{Is tmbfit}
\usage{
is.tmbfit(x)
}
\arguments{
\item{x}{The object.}
}
\value{
A flag.
}
\description{
Tests whether an object is a tmbfit object.
}
\seealso{
Other is: 
\code{\link{is.fitdistcens}()},
\code{\link{is.fitdistscens}()},
\code{\link{is.fitdists}()},
\code{\link{is.fitdist}()}
}
\concept{is}
