% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censor.R
\name{is_censored}
\alias{is_censored}
\title{Is Censored}
\usage{
is_censored(x)
}
\arguments{
\item{x}{A fitdists object.}
}
\value{
A flag indicating if the data is censored.
}
\description{
Is Censored
}
\examples{
boron_dists <- ssd_fit_dists(ssdtools::boron_data)
is_censored(boron_dists)
}
