% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensing.R
\name{licensing_md}
\alias{licensing_md}
\title{Licensing Markdown}
\usage{
licensing_md()
}
\description{
A string of markdown code indicating the licensing of the code
and documentation
}
\examples{
licensing_md()
}
