% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nobs.R
\name{nobs}
\alias{nobs}
\alias{nobs.fitdist}
\alias{nobs.fitdistcens}
\alias{nobs.fitdists}
\title{Number of Observations}
\usage{
nobs(object, ...)

\method{nobs}{fitdist}(object, ...)

\method{nobs}{fitdistcens}(object, ...)

\method{nobs}{fitdists}(object, ...)
}
\arguments{
\item{object}{The object.}

\item{...}{Unused.}
}
\description{
Number of Observations
}
\section{Methods (by class)}{
\itemize{
\item \code{fitdist}: Number of Observations fitdist

\item \code{fitdistcens}: Number of Observations fitdistcens

\item \code{fitdists}: Number of Observations fitdists
}}

\examples{
stats::nobs(boron_lnorm)
stats::nobs(fluazinam_lnorm)
}
\seealso{
\code{\link[stats:nobs]{stats::nobs()}}
}
