% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nobs.R
\name{nobs.tmbfit}
\alias{nobs.tmbfit}
\title{Number of Observations}
\usage{
\method{nobs}{tmbfit}(object, ...)
}
\arguments{
\item{object}{The object.}

\item{...}{Unused.}
}
\description{
Number of Observations
}
\seealso{
\code{\link[stats:nobs]{stats::nobs()}}
}
