% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pearson1000}
\alias{pearson1000}
\title{Pearson 1000 Data}
\format{
A tbl data frame that includes:
\describe{
\item{Conc}{A numeric vector of the simulate concentrations.}
}
}
\usage{
pearson1000
}
\description{
An example tibble of 1000 values simulated using a Pearson distribution
with a #FIXME of #FIXME and a #FIXME of #FIXME.
}
\details{
The data is released under $FIXME
}
\examples{
head(pearson1000)
}
\keyword{datasets}
