% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.fitdist}
\alias{predict.fitdist}
\title{Predict fitdist}
\usage{
\method{predict}{fitdist}(
  object,
  percent = 1:99,
  nboot = 1000,
  level = 0.95,
  parallel = "no",
  ncpus = 1,
  ...
)
}
\arguments{
\item{object}{The object.}

\item{percent}{A numeric vector of the densities to calculate the estimated concentrations for.}

\item{nboot}{A count of the number of bootstrap samples to use to estimate the se and confidence limits.}

\item{level}{The confidence level.}

\item{parallel}{A string of "no", "snow" or "multicore".
The "multicore" option is not avaiable on Windows.}

\item{ncpus}{The number of parallel processes to use.
Recommended to be the number of available CPUs.}

\item{...}{Unused}
}
\description{
Predict fitdist
}
\examples{
predict(boron_lnorm, percent = c(5L, 50L))
}
