% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.fitdists}
\alias{predict.fitdists}
\title{Predict fitdists}
\usage{
\method{predict}{fitdists}(
  object,
  percent = 1:99,
  nboot = 1000,
  ic = "aicc",
  average = TRUE,
  level = 0.95,
  parallel = "no",
  ncpus = 1L,
  ...
)
}
\arguments{
\item{object}{The object.}

\item{percent}{A numeric vector of the densities to calculate the estimated concentrations for.}

\item{nboot}{A count of the number of bootstrap samples to use to estimate the se and confidence limits.}

\item{ic}{A string specifying which information-theoretic criterion ('aic', 'aicc' or 'bic') to use for model averaging .}

\item{average}{A flag indicating whether to model-average.}

\item{level}{A number between 0 and 1 of the confidence level.}

\item{parallel}{A string specifying the type of parallel operation to be used ('no', 'snow' or 'multicore').}

\item{ncpus}{A count of the number of parallel processes to use.}

\item{...}{Unused.}
}
\description{
Predict fitdists
}
\examples{
\dontrun{
predict(boron_dists)
}
}
