% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.fitdistscens}
\alias{predict.fitdistscens}
\title{Predict Censored fitdists}
\usage{
\method{predict}{fitdistscens}(
  object,
  percent = 1:99,
  nboot = 1000,
  ic = "aic",
  average = TRUE,
  level = 0.95,
  parallel = "no",
  ncpus = ncpus,
  ...
)
}
\arguments{
\item{object}{The object.}

\item{percent}{A numeric vector of the densities to calculate the estimated concentrations for.}

\item{nboot}{A count of the number of bootstrap samples to use to estimate the se and confidence limits.}

\item{ic}{A string indicating which information-theoretic criterion ('aic', 'aicc' or 'bic') to use for model averaging.}

\item{average}{A flag indicating whether to model-average.}

\item{level}{The confidence level.}

\item{parallel}{A string of "no", "snow" or "multicore".
The "multicore" option is not avaiable on Windows.}

\item{ncpus}{The number of parallel processes to use.
Recommended to be the number of available CPUs.}

\item{...}{Unused}
}
\description{
Predict Censored fitdists
}
\examples{
\dontrun{
predict(fluazinam_dists)
}
}
