% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgumbel.R
\name{qlgumbel}
\alias{qlgumbel}
\title{Quantile Function for Log-Gumbel Distribution
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
qlgumbel(p, locationlog = 0, scalelog = 1, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{p}{vector of probabilities.}

\item{locationlog}{location on the log scale parameter.}

\item{scalelog}{scale on log scale parameter.}

\item{lower.tail}{logical; if TRUE (default), probabilities are \code{P[X <= x]}, otherwise, \code{P[X > x]}.}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}
}
\description{
Deprecated for \code{ssd_qlgumbel()}.
}
\keyword{internal}
