% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R, R/autoplot.R, R/coef.R,
%   R/estimates.R, R/glance.R, R/logLik.R, R/nobs.R, R/npars.R, R/plot.R,
%   R/predict.R, R/ssd-plot.R, R/tidy.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{augment}
\alias{autoplot}
\alias{coef}
\alias{estimates}
\alias{glance}
\alias{logLik}
\alias{nobs}
\alias{npars}
\alias{plot}
\alias{predict}
\alias{waiver}
\alias{tidy}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics:augment]{augment()}}, \code{\link[generics:glance]{glance()}}, \code{\link[generics:tidy]{tidy()}}}

  \item{ggplot2}{\code{\link[ggplot2:autoplot]{autoplot()}}, \code{\link[ggplot2:waiver]{waiver()}}}

  \item{graphics}{\code{\link[graphics:plot]{plot}}}

  \item{stats}{\code{\link[stats:coef]{coef()}}, \code{\link[stats:logLik]{logLik()}}, \code{\link[stats:nobs]{nobs()}}, \code{\link[stats:predict]{predict()}}}

  \item{universals}{\code{\link[universals:estimates]{estimates()}}, \code{\link[universals:npars]{npars()}}}
}}

