% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompertz.R
\name{rgompertz}
\alias{rgompertz}
\title{Random Generation for Gompertz Distribution
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
rgompertz(n, llocation = 0, lshape = 0)
}
\arguments{
\item{n}{positive number of observations.}

\item{llocation}{location parameter on the log scale.}

\item{lshape}{shape parameter on the log scale.}
}
\description{
Deprecated for \code{ssd_rgompertz()}.
}
