% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgumbel.R
\name{rlgumbel}
\alias{rlgumbel}
\title{Random Generation for log-Gumbel Distribution}
\usage{
rlgumbel(n, locationlog = 0, scalelog = 1)
}
\arguments{
\item{n}{positive number of observations.}

\item{locationlog}{location on the log scale parameter.}

\item{scalelog}{scale on log scale parameter.}
}
\description{
Deprecated for \code{ssd_rlgumbel()}.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\keyword{internal}
