% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ssd_cfplot}
\alias{ssd_cfplot}
\title{Cullen and Frey Plot}
\usage{
ssd_cfplot(data, left = "Conc")
}
\arguments{
\item{data}{A data frame.}

\item{left}{A string of the column in data with the left concentration values.}
}
\description{
Plots a Cullen and Frey graph of the skewness and kurtosis
for non-censored data.
}
\examples{
ssd_cfplot(boron_data)
}
\seealso{
\code{\link[fitdistrplus]{descdist}}
}
