% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci-methods.R
\name{ssd_ci_methods}
\alias{ssd_ci_methods}
\title{Confidence Interval Methods for SSDs}
\usage{
ssd_ci_methods()
}
\value{
A character vector of the available methods.
}
\description{
Returns a character vector of the available non-deprecated
methods for getting the model averaged confidence limits for two or more
distributions.
}
\examples{
ssd_ci_methods()
}
\keyword{internal}
