% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ssd_dists}
\alias{ssd_dists}
\title{Distribution Names}
\usage{
ssd_dists(all = FALSE)
}
\arguments{
\item{all}{A flag indicating whether to return all the distribution names.}
}
\value{
A sorted character vector of the distribution names.
}
\description{
Returns a sorted character vector of the recognized distribution names.
}
\examples{
ssd_dists()
ssd_dists(all = TRUE)
}
