% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burrrIII3.R, R/gamma.R, R/gompertz.R,
%   R/invpareto.R, R/lgumbel.R, R/llogis-llogis.R, R/llogis.R, R/lnorm-lnorm.R,
%   R/lnorm.R, R/ltriangle.R, R/ltriangle3.R, R/multi.R, R/pqr.R, R/weibull.R
\name{ssd_eburrIII3}
\alias{ssd_eburrIII3}
\alias{ssd_egamma}
\alias{ssd_egompertz}
\alias{ssd_einvpareto}
\alias{ssd_elgumbel}
\alias{ssd_ellogis_llogis}
\alias{ssd_ellogis}
\alias{ssd_elnorm_lnorm}
\alias{ssd_elnorm}
\alias{ssd_eltriangle}
\alias{ssd_eltriangle3}
\alias{ssd_emulti}
\alias{ssd_e}
\alias{ssd_eweibull}
\title{Default Parameter Estimates}
\usage{
ssd_eburrIII3()

ssd_egamma()

ssd_egompertz()

ssd_einvpareto()

ssd_elgumbel()

ssd_ellogis_llogis()

ssd_ellogis()

ssd_elnorm_lnorm()

ssd_elnorm()

ssd_eltriangle()

ssd_eltriangle3()

ssd_emulti()

ssd_eweibull()
}
\description{
Default Parameter Estimates
}
\section{Functions}{
\itemize{
\item \code{ssd_eburrIII3()}: Default Parameter Values for BurrIII Distribution

\item \code{ssd_egamma()}: Default Parameter Values for Gamma Distribution

\item \code{ssd_egompertz()}: Default Parameter Values for Gompertz Distribution

\item \code{ssd_einvpareto()}: Default Parameter Values for Inverse Pareto Distribution

\item \code{ssd_elgumbel()}: Default Parameter Values for log-Gumbel Distribution

\item \code{ssd_ellogis_llogis()}: Default Parameter Values for Log-Logistic/Log-Logistic Mixture Distribution

\item \code{ssd_ellogis()}: Default Parameter Values for Log-Logistic Distribution

\item \code{ssd_elnorm_lnorm()}: Default Parameter Values for Log-Normal/Log-Normal Mixture Distribution

\item \code{ssd_elnorm()}: Default Parameter Values for Log-Normal Distribution

\item \code{ssd_eltriangle()}: Default Parameter Values for Log-Triangular Distribution

\item \code{ssd_eltriangle3()}: Default Parameter Values for Log-Asymmetric-Triangular Distribution

\item \code{ssd_emulti()}: Default Parameter Values for Multiple Distributions

\item \code{ssd_eweibull()}: Default Parameter Values for Log-Normal Distribution

}}
\examples{

ssd_eburrIII3()

ssd_egamma()

ssd_egompertz()

ssd_einvpareto()

ssd_elgumbel()

ssd_ellogis_llogis()

ssd_ellogis()

ssd_elnorm_lnorm()

ssd_elnorm()

ssd_eltriangle()

ssd_eltriangle3()

ssd_emulti()

ssd_eweibull()
}
\seealso{
\code{\link{ssd_p}} and \code{\link{ssd_q}}
}
