% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ssd_ecd}
\alias{ssd_ecd}
\title{Empirical Cumulative Density
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
ssd_ecd(x, ties.method = "first")
}
\arguments{
\item{x}{a numeric, complex, character or logical vector.}

\item{ties.method}{a character string specifying how ties are treated,
    see \sQuote{Details}; can be abbreviated.}
}
\value{
A numeric vector of the empirical cumulative distribution function.
}
\description{
Deprecated for \code{\link[=ssd_ecdf]{ssd_ecdf()}}.
}
\examples{
\dontrun{
ssd_ecd(1:10)
}
}
\seealso{
\code{\link[=ssd_ecdf]{ssd_ecdf()}}
}
\keyword{internal}
