% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ssd_ecdf}
\alias{ssd_ecdf}
\title{Empirical Cumulative Distribution Function}
\usage{
ssd_ecdf(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector of the empirical cumulative distribution function.
}
\description{
Calculates the empirical cumulative distribution function of a numeric
vector using plotting positions, with ties ranked in order of occurrence.
}
\examples{
ssd_ecdf(1:10)
}
