% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element.R
\name{ssd_element_text_hc}
\alias{ssd_element_text_hc}
\title{Theme Element for Hazard Concentration Axis Labels}
\usage{
ssd_element_text_hc(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[ggplot2:element_text]{ggplot2::element_text()}}.}
}
\value{
An object of class \code{element_text_hc}.
}
\description{
A \code{\link[ggplot2:element_text]{ggplot2::element_text()}} that draws multi-line labels in bold.
It is required to make the hazard concentration label produced by
\code{\link[=ssd_label_comma_hc]{ssd_label_comma_hc()}} bold.
}
\examples{
ggplot2::ggplot(data = ssddata::anon_e, ggplot2::aes(x = Conc / 10)) +
  geom_ssdpoint() +
  ggplot2::scale_x_log10(labels = ssd_label_comma_hc(1.26)) +
  ggplot2::theme(axis.text.x = ssd_element_text_hc())
}
\seealso{
\code{\link[=ssd_label_comma_hc]{ssd_label_comma_hc()}}
}
