% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est-methods.R
\name{ssd_est_methods}
\alias{ssd_est_methods}
\title{Estimate Methods for SSDs}
\usage{
ssd_est_methods()
}
\value{
A character vector of the available methods.
}
\description{
Returns a character vector of the available non-deprecated
methods for getting the model averaged estimates for two or more
distributions.
}
\examples{
ssd_est_methods()
}
