% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcanz.R
\name{ssd_fit_bcanz}
\alias{ssd_fit_bcanz}
\title{Fit BCANZ Distributions}
\usage{
ssd_fit_bcanz(
  data,
  left = "Conc",
  ...,
  dists = ssd_dists_bcanz(),
  rescale = FALSE,
  silent = FALSE
)
}
\arguments{
\item{data}{A data frame.}

\item{left}{A string of the column in data with the concentrations.}

\item{...}{Unused.}

\item{dists}{A character vector of the distribution names.}

\item{rescale}{A flag specifying whether to leave the values unchanged (FALSE) or to rescale concentration values by dividing by the geometric mean of the minimum and maximum positive finite values (TRUE).}

\item{silent}{A flag indicating whether fits should fail silently.}
}
\value{
An object of class fitdists.
}
\description{
Fits distributions using settings adopted by
BC, Canada, Australia and New Zealand for official guidelines.
}
\examples{
ssd_fit_bcanz(ssddata::ccme_boron)
}
\seealso{
\code{\link[=ssd_fit_dists]{ssd_fit_dists()}}

Other BCANZ: 
\code{\link{ssd_hc_bcanz}()},
\code{\link{ssd_hp_bcanz}()}
}
\concept{BCANZ}
