% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-dist.R
\name{ssd_fit_dist}
\alias{ssd_fit_dist}
\title{Fit Distribution}
\usage{
ssd_fit_dist(x, dist = "lnorm")
}
\arguments{
\item{x}{A numeric vector of the data.}

\item{dist}{A string of the distribution to fit.}
}
\value{
An object of class fitdist.
}
\description{
Fit Distribution
}
\examples{
ssd_fit_dist(ccme_data$Conc[ccme_data$Chemical == "Boron"])
}
