% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jurisdictions.R
\name{ssd_fit_dists_bc}
\alias{ssd_fit_dists_bc}
\title{Fit Distributions for British Columbia}
\usage{
ssd_fit_dists_bc(data, left = "Conc", silent = FALSE)
}
\arguments{
\item{data}{A data frame.}

\item{left}{A string of the column in data with the concentrations.}

\item{silent}{A flag indicating whether fits should fail silently.}
}
\value{
An object of class fitdists.
}
\description{
Fits the gamma, llogis and lnorm distributions to uncensored,
non-weighted data that is automatically rescaled.
}
\details{
For more control see \code{\link[=ssd_fit_dists]{ssd_fit_dists()}}.
}
\examples{
ssd_fit_dists_bc(boron_data)
}
\seealso{
\code{\link[=ssd_hc5_bc]{ssd_hc5_bc()}}
}
