% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc.R
\name{ssd_hc}
\alias{ssd_hc}
\alias{ssd_hc.list}
\alias{ssd_hc.fitdists}
\alias{ssd_hc.fitburrlioz}
\title{Hazard Concentrations for Species Sensitivity Distributions}
\usage{
ssd_hc(x, ...)

\method{ssd_hc}{list}(x, percent = 5, ...)

\method{ssd_hc}{fitdists}(
  x,
  percent = 5,
  ci = FALSE,
  level = 0.95,
  nboot = 1000,
  average = TRUE,
  multi_est = TRUE,
  delta = 7,
  min_pboot = 0.99,
  parametric = TRUE,
  multi_ci = TRUE,
  weighted = TRUE,
  control = NULL,
  samples = FALSE,
  save_to = NULL,
  ...
)

\method{ssd_hc}{fitburrlioz}(
  x,
  percent = 5,
  ci = FALSE,
  level = 0.95,
  nboot = 1000,
  min_pboot = 0.99,
  parametric = FALSE,
  save_to = NULL,
  samples = FALSE,
  ...
)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused.}

\item{percent}{A numeric vector of percent values to estimate hazard concentrations for.}

\item{ci}{A flag specifying whether to estimate confidence intervals (by bootstrapping).}

\item{level}{A number between 0 and 1 of the confidence level.}

\item{nboot}{A count of the number of bootstrap samples to use to estimate confidence limits. A value of at least 10,000 is recommended for official guidelines.}

\item{average}{A flag specifying whether to provide model averaged values.}

\item{multi_est}{A flag specifying whether to treat the distributions as constituting a single distribution when providing model averaged estimates.}

\item{delta}{A non-negative number specifying the maximum absolute AIC difference cutoff.
Distributions with an absolute AIC difference greater than delta are excluded.}

\item{min_pboot}{A number between 0 and 1 of the minimum
proportion of bootstrap samples that must successfully fit in the sense of returning a likelihood
in order to report the confidence intervals.}

\item{parametric}{A flag specifying whether to perform parametric (as opposed to non-parametric) bootstrapping when \code{ci = TRUE}.}

\item{multi_ci}{A flag specifying whether to treat the distributions as constituting a single distribution when generating model averaged confidence intervals.}

\item{weighted}{A flag which when \code{multi_ci = TRUE} specifies whether to use the original model weights (as opposed to reestimating for each bootstrap sample) or when \code{multi_ci = FALSE} specifies whether to use the weighted bootstrap (as opposed to generating the same number of bootstrap samples for each model).}

\item{control}{A list of control parameters passed to \code{\link[stats:optim]{stats::optim()}}.}

\item{samples}{A flag specfying whether to include a numeric vector of the bootstrap samples as a list column in the output.}

\item{save_to}{NULL or a string specifying a directory to save the bootstrap datasets and parameter estimates to.}
}
\value{
A tibble of corresponding hazard concentrations.
}
\description{
Calculates concentration(s) with quantile based bootstrap confidence intervals
that protect specified percentage(s) of species for
individual or model-averaged distributions
using parametric or non-parametric bootstrapping.
}
\details{
Model-averaged estimates and/or confidence intervals (including standard error)
can be calculated  by treating the distributions as constituting a single mixture distribution
versus 'taking the mean'.

If treating the distributions as constituting a single mixture distribution
when calculating model average confidence intervals then
\code{weighted} specifies whether to use the original model weights versus
re-estimating for each bootstrap sample.
Conversely when 'taking the mean' then \code{weighted} specifies
whether to take bootstrap samples from each distribution proportional to
its weight versus calculating the weighted arithmetic means of the lower
and upper confidence limits.

Based on Burnham and Anderson (2002),
distributions with an absolute AIC difference greater
than a delta of by default 7 are excluded
prior to calculation of the hazard concentrations.
}
\section{Methods (by class)}{
\itemize{
\item \code{ssd_hc(list)}: Hazard Concentrations for Distributional Estimates

\item \code{ssd_hc(fitdists)}: Hazard Concentrations for fitdists Object

\item \code{ssd_hc(fitburrlioz)}: Hazard Concentrations for fitburrlioz Object

}}
\examples{

ssd_hc(ssd_match_moments())

fits <- ssd_fit_dists(ssddata::ccme_boron)
ssd_hc(fits)

fit <- ssd_fit_burrlioz(ssddata::ccme_boron)
ssd_hc(fit)
}
\references{
Burnham, K.P., and Anderson, D.R. 2002. Model Selection and Multimodel Inference. Springer New York, New York, NY. doi:10.1007/b97636.
}
\seealso{
\code{\link[=predict.fitdists]{predict.fitdists()}} and \code{\link[=ssd_hp]{ssd_hp()}}.
}
