% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc5.R
\name{ssd_hc5.fitdistcens}
\alias{ssd_hc5.fitdistcens}
\title{5\% Hazard Concentration}
\usage{
\method{ssd_hc5}{fitdistcens}(x, level = 0.95, ...)
}
\arguments{
\item{x}{A fitdistcens object.}

\item{level}{The confidence level.}

\item{...}{Unused.}
}
\description{
Estimates with bootstrap confidence intervals the hazard concentration
at which 5\% of the species are affected.
}
\examples{
ssd_hc5(fluazinam_lnorm)
}
