% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc5.R
\name{ssd_hc5.fitdistscens}
\alias{ssd_hc5.fitdistscens}
\title{5\% Hazard Concentration}
\usage{
\method{ssd_hc5}{fitdistscens}(x, ic = "aic", average = TRUE,
  level = 0.95, ...)
}
\arguments{
\item{x}{A fitdistcens object.}

\item{ic}{A string indicating which information-theoretic criterion ('aic', 'aicc' or 'bic') to use for model averaging.}

\item{average}{A flag indicating whether to model-average.}

\item{level}{The confidence level.}

\item{...}{Unused.}
}
\description{
Estimates with bootstrap confidence intervals the hazard concentration
at which 5\% of the species are affected.
}
\examples{
\dontrun{
ssd_hc5(fluazinam_dists)
}
}
