% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jurisdictions.R
\name{ssd_hc5_bc}
\alias{ssd_hc5_bc}
\title{Hazard Concentration 5\% for British Columbia}
\usage{
ssd_hc5_bc(x)
}
\arguments{
\item{x}{A fitdists object fitted using \code{\link[=ssd_fit_dists_bc]{ssd_fit_dists_bc()}}}
}
\value{
A tbl data frame of the HC5.
}
\description{
Calculates the model averaged 5\% Hazard Concentration (HC5) with
95\% parametric bootstrap confidence intervals estimated from 10,000
samples for the gamma, llogis and lnorm distributions fitted to
uncensored, non-weighted, rescaled data.
}
\examples{
\dontrun{
fits <- ssd_fit_dists_bc(boron_data)
ssd_hc5_bc(fits)
}
}
\seealso{
\code{\link[=ssd_hc5_bc]{ssd_hc5_bc()}}
}
