% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc-burrlioz.R
\name{ssd_hc_burrlioz}
\alias{ssd_hc_burrlioz}
\title{Hazard Concentrations for Burrlioz Fit}
\usage{
ssd_hc_burrlioz(
  x,
  percent = 5,
  ci = FALSE,
  level = 0.95,
  nboot = 1000,
  min_pboot = 0.99
)
}
\arguments{
\item{x}{The object.}

\item{percent}{A numeric vector of percentages.}

\item{ci}{A flag specifying whether to estimate confidence intervals (by parametric bootstrapping).}

\item{level}{A number between 0 and 1 of the confidence level.}

\item{nboot}{A count of the number of bootstrap samples to use to estimate the se and confidence limits. A value of 10000 is recommended for official guidelines.}

\item{min_pboot}{A number of the minimum proportion of bootstrap samples that must successfully fit
in the sense of returning a likelihood.}
}
\value{
A tibble of corresponding hazard concentrations.
}
\description{
Gets concentration(s) that protect specified percentage(s) of species
based on the distribution.
}
\details{
If \code{ci = TRUE} uses non-parameteric bootstrapping to get confidence intervals on the
hazard concentrations(s).
}
\examples{
fit <- ssd_fit_burrlioz(ssddata::ccme_boron)
ssd_hc_burrlioz(fit)

}
\seealso{
\code{\link[=ssd_hc]{ssd_hc()}}
}
