% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min-pmix.R
\name{ssd_min_pmix}
\alias{ssd_min_pmix}
\title{Calculate Minimum Proportion in Mixture Models}
\usage{
ssd_min_pmix(n)
}
\arguments{
\item{n}{positive number of observations.}
}
\value{
A number between 0 and 0.5 of the minimum proportion in mixture models.
}
\description{
Calculate Minimum Proportion in Mixture Models
}
\examples{
ssd_min_pmix(6)
ssd_min_pmix(50)
}
\seealso{
\code{\link[=ssd_fit_dists]{ssd_fit_dists()}}
}
