% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{ssd_moments}
\alias{ssd_moments}
\title{Calculate Moments}
\usage{
ssd_moments(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A tbl data frame of the min, max, median, mean and unbiased estimates of the first four moments.
}
\description{
A wrapper on \code{\link[fitdistrplus]{descdist}}.
}
\examples{
ssd_moments(ccme_data$Conc)
}
