% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burrrIII3.R, R/gamma.R, R/gompertz.R,
%   R/invpareto.R, R/lgumbel.R, R/llogis-llogis.R, R/llogis.R, R/lnorm-lnorm.R,
%   R/lnorm.R, R/ltriangle.R, R/ltriangle3.R, R/multi.R, R/pqr.R, R/weibull.R
\name{ssd_pburrIII3}
\alias{ssd_pburrIII3}
\alias{ssd_pgamma}
\alias{ssd_pgompertz}
\alias{ssd_pinvpareto}
\alias{ssd_plgumbel}
\alias{ssd_pllogis_llogis}
\alias{ssd_pllogis}
\alias{ssd_plnorm_lnorm}
\alias{ssd_plnorm}
\alias{ssd_pltriangle}
\alias{ssd_pltriangle3}
\alias{ssd_pmulti}
\alias{ssd_pmulti_fitdists}
\alias{ssd_p}
\alias{ssd_pweibull}
\title{Cumulative Distribution Function}
\usage{
ssd_pburrIII3(
  q,
  shape1 = 1,
  shape2 = 1,
  scale = 1,
  lower.tail = TRUE,
  log.p = FALSE
)

ssd_pgamma(q, shape = 1, scale = 1, lower.tail = TRUE, log.p = FALSE)

ssd_pgompertz(q, location = 1, shape = 1, lower.tail = TRUE, log.p = FALSE)

ssd_pinvpareto(q, shape = 3, scale = 1, lower.tail = TRUE, log.p = FALSE)

ssd_plgumbel(
  q,
  locationlog = 0,
  scalelog = 1,
  lower.tail = TRUE,
  log.p = FALSE
)

ssd_pllogis_llogis(
  q,
  locationlog1 = 0,
  scalelog1 = 1,
  locationlog2 = 1,
  scalelog2 = 1,
  pmix = 0.5,
  lower.tail = TRUE,
  log.p = FALSE
)

ssd_pllogis(q, locationlog = 0, scalelog = 1, lower.tail = TRUE, log.p = FALSE)

ssd_plnorm_lnorm(
  q,
  meanlog1 = 0,
  sdlog1 = 1,
  meanlog2 = 1,
  sdlog2 = 1,
  pmix = 0.5,
  lower.tail = TRUE,
  log.p = FALSE
)

ssd_plnorm(q, meanlog = 0, sdlog = 1, lower.tail = TRUE, log.p = FALSE)

ssd_pltriangle(
  q,
  locationlog = 0,
  scalelog = 3,
  lower.tail = TRUE,
  log.p = FALSE
)

ssd_pltriangle3(
  q,
  locationlog = 0,
  scalelog = 3,
  skew = 0,
  lower.tail = TRUE,
  log.p = FALSE
)

ssd_pmulti(
  q,
  burrIII3.weight = 0,
  burrIII3.shape1 = 1,
  burrIII3.shape2 = 1,
  burrIII3.scale = 1,
  gamma.weight = 0,
  gamma.shape = 1,
  gamma.scale = 1,
  gompertz.weight = 0,
  gompertz.location = 1,
  gompertz.shape = 1,
  lgumbel.weight = 0,
  lgumbel.locationlog = 0,
  lgumbel.scalelog = 1,
  llogis.weight = 0,
  llogis.locationlog = 0,
  llogis.scalelog = 1,
  llogis_llogis.weight = 0,
  llogis_llogis.locationlog1 = 0,
  llogis_llogis.scalelog1 = 1,
  llogis_llogis.locationlog2 = 1,
  llogis_llogis.scalelog2 = 1,
  llogis_llogis.pmix = 0.5,
  lnorm.weight = 0,
  lnorm.meanlog = 0,
  lnorm.sdlog = 1,
  lnorm_lnorm.weight = 0,
  lnorm_lnorm.meanlog1 = 0,
  lnorm_lnorm.sdlog1 = 1,
  lnorm_lnorm.meanlog2 = 1,
  lnorm_lnorm.sdlog2 = 1,
  lnorm_lnorm.pmix = 0.5,
  ltriangle.weight = 0,
  ltriangle.locationlog = 0,
  ltriangle.scalelog = 3,
  ltriangle3.weight = 0,
  ltriangle3.locationlog = 0,
  ltriangle3.scalelog = 3,
  ltriangle3.skew = 0,
  weibull.weight = 0,
  weibull.shape = 1,
  weibull.scale = 1,
  lower.tail = TRUE,
  log.p = FALSE
)

ssd_pmulti_fitdists(q, fitdists, lower.tail = TRUE, log.p = FALSE)

ssd_pweibull(q, shape = 1, scale = 1, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{q}{vector of quantiles.}

\item{shape1}{shape1 parameter.}

\item{shape2}{shape2 parameter.}

\item{scale}{scale parameter.}

\item{lower.tail}{logical; if TRUE (default), probabilities are \code{P[X <= x]}, otherwise, \code{P[X > x]}.}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{shape}{shape parameter.}

\item{location}{location parameter.}

\item{locationlog}{location on the log scale parameter.}

\item{scalelog}{scale on log scale parameter.}

\item{locationlog1}{locationlog1 parameter.}

\item{scalelog1}{scalelog1 parameter.}

\item{locationlog2}{locationlog2 parameter.}

\item{scalelog2}{scalelog2 parameter.}

\item{pmix}{Proportion mixture parameter.}

\item{meanlog1}{mean on log scale parameter.}

\item{sdlog1}{standard deviation on log scale parameter.}

\item{meanlog2}{mean on log scale parameter.}

\item{sdlog2}{standard deviation on log scale parameter.}

\item{meanlog}{mean on log scale parameter.}

\item{sdlog}{standard deviation on log scale parameter.}

\item{skew}{skewness parameter between -1 and 1 (0 is symmetric).}

\item{burrIII3.weight}{weight parameter for the Burr III distribution.}

\item{burrIII3.shape1}{shape1 parameter for the Burr III distribution.}

\item{burrIII3.shape2}{shape2 parameter for the Burr III distribution.}

\item{burrIII3.scale}{scale parameter for the Burr III distribution.}

\item{gamma.weight}{weight parameter for the gamma distribution.}

\item{gamma.shape}{shape parameter for the gamma distribution.}

\item{gamma.scale}{scale parameter for the gamma distribution.}

\item{gompertz.weight}{weight parameter for the Gompertz distribution.}

\item{gompertz.location}{location parameter for the Gompertz distribution.}

\item{gompertz.shape}{shape parameter for the Gompertz distribution.}

\item{lgumbel.weight}{weight parameter for the log-Gumbel distribution.}

\item{lgumbel.locationlog}{location parameter for the log-Gumbel distribution.}

\item{lgumbel.scalelog}{scale parameter for the log-Gumbel distribution.}

\item{llogis.weight}{weight parameter for the log-logistic distribution.}

\item{llogis.locationlog}{location parameter for the log-logistic distribution.}

\item{llogis.scalelog}{scale parameter for the log-logistic distribution.}

\item{llogis_llogis.weight}{weight parameter for the log-logistic log-logistic mixture distribution.}

\item{llogis_llogis.locationlog1}{locationlog1 parameter for the log-logistic log-logistic mixture distribution.}

\item{llogis_llogis.scalelog1}{scalelog1 parameter for the log-logistic log-logistic mixture distribution.}

\item{llogis_llogis.locationlog2}{locationlog2 parameter for the log-logistic log-logistic mixture distribution.}

\item{llogis_llogis.scalelog2}{scalelog2 parameter for the log-logistic log-logistic mixture distribution.}

\item{llogis_llogis.pmix}{pmix parameter for the log-logistic log-logistic mixture distribution.}

\item{lnorm.weight}{weight parameter for the log-normal distribution.}

\item{lnorm.meanlog}{meanlog parameter for the log-normal distribution.}

\item{lnorm.sdlog}{sdlog parameter for the log-normal distribution.}

\item{lnorm_lnorm.weight}{weight parameter for the log-normal log-normal mixture distribution.}

\item{lnorm_lnorm.meanlog1}{meanlog1 parameter for the log-normal log-normal mixture distribution.}

\item{lnorm_lnorm.sdlog1}{sdlog1 parameter for the log-normal log-normal mixture distribution.}

\item{lnorm_lnorm.meanlog2}{meanlog2 parameter for the log-normal log-normal mixture distribution.}

\item{lnorm_lnorm.sdlog2}{sdlog2 parameter for the log-normal log-normal mixture distribution.}

\item{lnorm_lnorm.pmix}{pmix parameter for the log-normal log-normal mixture distribution.}

\item{ltriangle.weight}{weight parameter for the log-triangular distribution.}

\item{ltriangle.locationlog}{location parameter for the log-triangular distribution.}

\item{ltriangle.scalelog}{scale parameter for the log-triangular distribution.}

\item{ltriangle3.weight}{weight parameter for the log-asymmetric-triangular distribution.}

\item{ltriangle3.locationlog}{location parameter for the log-asymmetric-triangular distribution.}

\item{ltriangle3.scalelog}{scale parameter for the log-asymmetric-triangular distribution.}

\item{ltriangle3.skew}{skewness parameter for the log-asymmetric-triangular distribution.}

\item{weibull.weight}{weight parameter for the Weibull distribution.}

\item{weibull.shape}{shape parameter for the Weibull distribution.}

\item{weibull.scale}{scale parameter for the Weibull distribution.}

\item{fitdists}{An object of class fitdists.}
}
\description{
Cumulative Distribution Function
}
\section{Functions}{
\itemize{
\item \code{ssd_pburrIII3()}: Cumulative Distribution Function for BurrIII Distribution

\item \code{ssd_pgamma()}: Cumulative Distribution Function for Gamma Distribution

\item \code{ssd_pgompertz()}: Cumulative Distribution Function for Gompertz Distribution

\item \code{ssd_pinvpareto()}: Cumulative Distribution Function for Inverse Pareto Distribution

\item \code{ssd_plgumbel()}: Cumulative Distribution Function for Log-Gumbel Distribution

\item \code{ssd_pllogis_llogis()}: Cumulative Distribution Function for Log-Logistic/Log-Logistic Mixture Distribution

\item \code{ssd_pllogis()}: Cumulative Distribution Function for Log-Logistic Distribution

\item \code{ssd_plnorm_lnorm()}: Cumulative Distribution Function for Log-Normal/Log-Normal Mixture Distribution

\item \code{ssd_plnorm()}: Cumulative Distribution Function for Log-Normal Distribution

\item \code{ssd_pltriangle()}: Cumulative Distribution Function for Log-Triangular Distribution

\item \code{ssd_pltriangle3()}: Cumulative Distribution Function for Log-Asymmetric-Triangular Distribution

\item \code{ssd_pmulti()}: Cumulative Distribution Function for Multiple Distributions

\item \code{ssd_pmulti_fitdists()}: Cumulative Distribution Function for Multiple Distributions

\item \code{ssd_pweibull()}: Cumulative Distribution Function for Weibull Distribution

}}
\examples{

ssd_pburrIII3(1)

ssd_pgamma(1)

ssd_pgompertz(1)

ssd_pinvpareto(1)

ssd_plgumbel(1)

ssd_pllogis_llogis(1)

ssd_pllogis(1)

ssd_plnorm_lnorm(1)

ssd_plnorm(1)

ssd_pltriangle(1)

ssd_pltriangle3(1)

# multi
ssd_pmulti(1, gamma.weight = 0.5, lnorm.weight = 0.5)

# multi fitdists
fit <- ssd_fit_dists(ssddata::ccme_boron)
ssd_pmulti_fitdists(1, fit)

ssd_pweibull(1)
}
\seealso{
\code{\link{ssd_q}} and \code{\link{ssd_r}}
}
