% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{ssd_pal}
\alias{ssd_pal}
\title{Color-blind Palette for SSD Plots}
\usage{
ssd_pal()
}
\value{
A character vector of a color blind palette with 8 colors.
}
\description{
Color-blind Palette for SSD Plots
}
\examples{
ssd_pal()
}
\seealso{
Other ggplot: 
\code{\link{geom_hcintersect}()},
\code{\link{geom_ssdpoint}()},
\code{\link{geom_ssdsegment}()},
\code{\link{geom_xribbon}()},
\code{\link{scale_colour_ssd}()}
}
\concept{ggplot}
