% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-skewness-kurtosis.R
\name{ssd_plot_skewness_kurtosis}
\alias{ssd_plot_skewness_kurtosis}
\title{Plot Skewness-Kurtosis Graph}
\usage{
ssd_plot_skewness_kurtosis(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
An invisible TRUE.
}
\description{
Plots a skewness-kurtosis graph similar to the one proposed by Cullen and Frey with 100 bootstrapped values.
}
\examples{
ssd_plot_skewness_kurtosis(ccme_data$Conc[ccme_data$Chemical == "Cadmium"])
}
\seealso{
\code{\link[fitdistrplus]{descdist}}
}
