% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp.R
\name{ssd_pp}
\alias{ssd_pp}
\alias{ssd_pp.fitdist}
\title{Percent Protected}
\usage{
ssd_pp(x, ...)

\method{ssd_pp}{fitdist}(x, conc, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused.}

\item{conc}{A numeric vector of the concentrations.}
}
\value{
A numeric vector of percent values between 0 and 100.
}
\description{
Gets percent species protected at specified concentrations.
}
\section{Methods (by class)}{
\itemize{
\item \code{fitdist}: Percent Protected ssd_pp
}}

\examples{
ssd_pp(boron_lnorm, c(0, 1, 30, Inf))
}
