% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ssd_skplot}
\alias{ssd_skplot}
\title{Plot Skewness-Kurtosis Graph}
\usage{
ssd_skplot(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
An invisible TRUE.
}
\description{
Plots a skewness-kurtosis graph similar to the one proposed by Cullen and Frey with 100 bootstrapped values.
}
\examples{
ssd_skplot(ccme_data$Concentration[ccme_data$Chemical == "Cadmium"])
}
\seealso{
\code{\link[fitdistrplus]{descdist}}
}
