% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqg.R
\name{ssd_wqg_burrlioz}
\alias{ssd_wqg_burrlioz}
\title{Water Quality Guideline for Burrlioz
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
ssd_wqg_burrlioz(data, left = "Conc")
}
\arguments{
\item{data}{A data frame.}

\item{left}{A string of the column in data with the concentrations.}
}
\value{
A tibble of the 5\% hazard concentration with 95\% confidence intervals.
}
\description{
Calculates the 5\% Hazard Concentration using \code{ssd_fit_burrlioz()}
and \code{ssd_hc()}.
}
\examples{
\dontrun{
ssd_wqg_burrlioz(ssddata::ccme_boron)
}
}
\seealso{
\code{\link[=ssd_fit_burrlioz]{ssd_fit_burrlioz()}} and \code{\link[=ssd_hc]{ssd_hc()}}

Other wqg: 
\code{\link{ssd_wqg_bc}()}
}
\concept{wqg}
\keyword{internal}
