% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt-est.R
\name{ssd_wt_est}
\alias{ssd_wt_est}
\title{Get Distributions Weights and Parameter Estimates}
\usage{
ssd_wt_est(x)
}
\arguments{
\item{x}{The object.}
}
\description{
Gets distribution names, weights and parameter estimates for a fitdists object.
}
\examples{
fits <- ssd_fit_dists(ssddata::ccme_boron)
ssd_wt_est(fits)
ssd_wt_est(fits)$data[[1]]
}
