% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggproto.R
\name{ssdtools-ggproto}
\alias{ssdtools-ggproto}
\alias{StatSsdpoint}
\alias{StatSsdsegment}
\alias{GeomSsdpoint}
\alias{GeomSsdsegment}
\alias{GeomHcintersect}
\alias{GeomXribbon}
\title{ggproto Classes for Plotting Species Sensitivity Data and Distributions}
\usage{
StatSsdpoint

StatSsdsegment

GeomSsdpoint

GeomSsdsegment

GeomHcintersect

GeomXribbon
}
\description{
ggproto Classes for Plotting Species Sensitivity Data and Distributions
}
\seealso{
\code{\link[ggplot2:ggproto]{ggplot2::ggproto()}} and \code{\link[=ssd_plot_cdf]{ssd_plot_cdf()}}
}
