% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{Test Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 141 rows and 2 columns.
}
\usage{
test_data
}
\description{
Data to test ssdtools.
}
\details{
\describe{
\item{Chemical}{The chemical (chr).}
\item{Conc}{The chemical concentration (dbl).}
}
}
\examples{
head(test_data)
}
\keyword{datasets}
