% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.fitdists}
\alias{tidy.fitdists}
\title{Turn a fitdists object into a tidy tibble}
\usage{
\method{tidy}{fitdists}(x, all = FALSE, ...)
}
\arguments{
\item{x}{A fitdists object to be converted into a tidy tibble::tibble().}

\item{all}{A flag specifying whether to also return transformed parameters.}

\item{...}{Must be unused.}
}
\description{
Turns a fitdists object into a tidy tibble.
}
