% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.tmbfit}
\alias{tidy.tmbfit}
\title{Turn a tmbfit object into a tidy tibble}
\usage{
\method{tidy}{tmbfit}(x, all = FALSE, ...)
}
\arguments{
\item{x}{A tmbfit object to be converted into a tidy tibble::tibble().}

\item{all}{A flag specifying whether to also return transformed parameters.}

\item{...}{Unused.}
}
\description{
Turns a tmbfit object into a tidy tibble.
}
