% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{augment.fitdists}
\alias{augment.fitdists}
\title{Augment data with information from an object.}
\usage{
\method{augment}{fitdists}(x, ...)
}
\arguments{
\item{x}{A fitdists object.}

\item{...}{Unused.}
}
\description{
Augment data with information from an object.
}
