% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{augment.fitdists}
\alias{augment.fitdists}
\title{Augmented Data from fitdists Object}
\usage{
\method{augment}{fitdists}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused.}
}
\value{
A tibble of the agumented data.
}
\description{
Get a tibble of the original data with augmentation.
}
\examples{
fits <- ssd_fit_dists(ssdtools::boron_data)
augment(fits)
}
\seealso{
\code{\link[=ssd_data]{ssd_data()}}

Other generics: 
\code{\link{glance.fitdists}()},
\code{\link{tidy.fitdists}()}
}
\concept{generics}
