% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.fitdist}
\alias{autoplot.fitdist}
\title{Autoplot}
\usage{
\method{autoplot}{fitdist}(object, ci = FALSE, hc5 = TRUE,
  xlab = "Concentration", ylab = "Species Affected", ...)
}
\arguments{
\item{object}{The object to plot.}

\item{ci}{A flag indicating wether to plot confidence intervals}

\item{hc5}{A flag indicating whether to plot the HC5.}

\item{xlab}{A string of the x-axis label.}

\item{ylab}{A string of the x-axis label.}

\item{...}{Unused.}
}
\description{
Autoplot
}
\examples{
autoplot(boron_lnorm)
}
