% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.fitdistcens}
\alias{autoplot.fitdistcens}
\title{Autoplot}
\usage{
\method{autoplot}{fitdistcens}(object, ci = FALSE, hc = 5L,
  xlab = "Concentration", ylab = "Species Affected", ...)
}
\arguments{
\item{object}{The object to plot.}

\item{ci}{A flag indicating wether to plot confidence intervals}

\item{hc}{A count between 1 and 99 indicating the percent hazard concentration to plot (or NULL).}

\item{xlab}{A string of the x-axis label.}

\item{ylab}{A string of the x-axis label.}

\item{...}{Unused.}
}
\description{
Autoplot
}
\examples{
fluazinam_lnorm$censdata$right[3] <- fluazinam_lnorm$censdata$left[3] * 1.5
fluazinam_lnorm$censdata$left[5] <- NA
ggplot2::autoplot(fluazinam_lnorm)
}
