% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.fitdists}
\alias{autoplot.fitdists}
\title{Autoplot}
\usage{
\method{autoplot}{fitdists}(object, xlab = "Concentration",
  ylab = "Percent of Species Affected", ...)
}
\arguments{
\item{object}{The object to plot.}

\item{xlab}{A string of the x-axis label.}

\item{ylab}{A string of the x-axis label.}

\item{...}{Unused.}
}
\description{
Autoplot
}
\examples{
\dontrun{
library(ggplot2)
autoplot(boron_dists)
}
}
