% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boron_data}
\alias{boron_data}
\title{CCME Species Sensitivity Data for Boron}
\format{
A tbl data frame:
\describe{
  \item{Chemical}{The chemical <chr>.}
  \item{Species}{The species binomial name <chr>.}
  \item{Concentration}{The chemical concentration <dbl>.}
  \item{Units}{The units <chr>.}
  \item{Group}{The taxonomic group <fctr>.}
}
}
\usage{
boron_data
}
\description{
Species Sensitivity Data from
the Canadian Council of Ministers of the Environment.
}
\details{
Additional information is available from \url{http://ceqg-rcqe.ccme.ca/download/en/324/}.
}
\examples{
head(ccme_data)
}
\seealso{
ccme_data
}
\keyword{datasets}
