% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boron_hc5}
\alias{boron_hc5}
\title{Model averaged 5\% hazard concentration for CCME Boron Data}
\format{
A tbl data frame:
\describe{
  \item{percent}{The percent of species affected <int>.}
  \item{est}{The estimated concentration <dbl>.}
  \item{se}{The standard error of the estimate <dbl>.}
  \item{lcl}{The lower confidence limit <dbl>.}
  \item{se}{The upper confidence limit <dbl>.}
  \item{dist}{The distribution <chr>.}
}
}
\usage{
boron_hc5
}
\description{
A data frame of the predictions based on 10000 bootstrap.
}
\examples{
boron_hc5
}
\keyword{datasets}
