% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boron_pred}
\alias{boron_pred}
\title{Model averaged predictions for CCME Boron Data}
\format{A tbl data frame:
\describe{
  \item{percent}{The percent of species affected <int>.}
  \item{est}{The estimated concentration <dbl>.}
  \item{se}{The standard error of the estimate <dbl>.}
  \item{lcl}{The lower confidence limit <dbl>.}
  \item{se}{The upper confidence limit <dbl>.}
}}
\usage{
boron_pred
}
\description{
A data frame of the predictions.
}
\examples{
head(boron_pred)
}
\keyword{datasets}
