% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burrIII2.R
\name{burrIII2}
\alias{burrIII2}
\alias{dburrIII2}
\alias{pburrIII2}
\alias{qburrIII2}
\alias{rburrIII2}
\title{Burr Type III Two-Parameter Distribution}
\usage{
dburrIII2(x, locationlog = 0, scalelog = 1, log = FALSE)

pburrIII2(q, locationlog = 0, scalelog = 1, lower.tail = TRUE, log.p = FALSE)

qburrIII2(p, locationlog = 0, scalelog = 1, lower.tail = TRUE, log.p = FALSE)

rburrIII2(n, locationlog = 0, scalelog = 1)
}
\arguments{
\item{x}{A numeric vector of values.}

\item{locationlog}{location on log scale parameter.}

\item{scalelog}{scale on log scale parameter.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}

\item{q}{vector of quantiles.}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X <= x],otherwise, P[X > x].}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
A numeric vector.
}
\description{
Probability density, cumulative distribution, 
inverse cumulative distribution, random sample and starting values functions
for the
Burr Type III two-parameter distribution
with \code{locationlog} and \code{scalelog} parameters.
}
\examples{
x <- seq(0.01, 5, by = 0.01)
plot(x, dburrIII2(x), type = "l")
}
\seealso{
\code{\link{llogis}}
}
