% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burrIII2.R
\name{burrIII2}
\alias{burrIII2}
\alias{dburrIII2}
\alias{qburrIII2}
\alias{pburrIII2}
\alias{rburrIII2}
\title{Burr Type III Two-Parameter Distribution}
\usage{
dburrIII2(x, lshape = log(1), lscale = log(1), log = FALSE)

qburrIII2(
  p,
  lshape = log(1),
  lscale = log(1),
  lower.tail = TRUE,
  log.p = FALSE
)

pburrIII2(
  q,
  lshape = log(1),
  lscale = log(1),
  lower.tail = TRUE,
  log.p = FALSE
)

rburrIII2(n, lshape = log(1), lscale = log(1))
}
\arguments{
\item{x}{The object.}

\item{lshape}{log shape parameter.}

\item{lscale}{log scale parameter.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}

\item{p}{vector of probabilities.}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X <= x],otherwise, P[X > x].}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{q}{vector of quantiles.}

\item{n}{number of observations.}
}
\value{
dburrIII2 gives the density, pburrIII2 gives the distribution function,
qburrIII2 gives the quantile function, and rburrIII2 generates random samples.
}
\description{
Density, distribution function, quantile function and random generation
for the Burr Type III Two-Parameter distribution with 
\code{lshape} and \code{lscale} parameters.
}
\details{
The Burr 12 distribution from the actuar package is used as a base.
The Burr III distribution is the distribution of 1/x where x has the Burr Type 12 distribution.
refer to https://www.itl.nist.gov/div898/software/dataplot/refman2/auxillar/bu3pdf.htm for details.
The shape1, shape2, and scale paramters are on the log(scale) as these must be positive.
The two shape parameters are set to be equal to produce a two-parameter model.
}
\examples{
x <- rburrIII2(1000)
hist(x, freq = FALSE, col = "gray", border = "white")
curve(dburrIII2(x), add = TRUE, col = "red4", lwd = 2)
}
\seealso{
\code{\link[actuar]{dburr}}
}
