% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{comma_signif}
\alias{comma_signif}
\title{Comma and Significance Formatter}
\usage{
comma_signif(x, digits = 1, ...)
}
\arguments{
\item{x}{A numeric vector to format.}

\item{digits}{Deprecated, use \code{accuracy} instead.}

\item{...}{Other arguments passed on to \code{\link[base:format]{base::format()}}.}
}
\value{
A function that returns a character vector.
}
\description{
Comma and Significance Formatter
}
\examples{
comma_signif(1199)
}
\seealso{
\code{\link[scales:label_number]{scales::comma()}}
}
