% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{comma_signif}
\alias{comma_signif}
\title{Comma and Significance Formatter}
\usage{
comma_signif(x, digits = 3, ..., big.mark = ",")
}
\arguments{
\item{x}{A numeric vector to format.}

\item{digits}{A whole number specifying the number of significant figures.}

\item{...}{Unused.}

\item{big.mark}{A string specifying used between every 3 digits to separate thousands.}
}
\value{
A character vector.
}
\description{
Comma and Significance Formatter
}
\examples{
comma_signif(c(0.1, 1, 10, 1000, 10000))
}
