% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{comma_signif}
\alias{comma_signif}
\title{Comma and Significance Formatter}
\usage{
comma_signif(x, digits = 1, ...)
}
\arguments{
\item{x}{a numeric vector to format}

\item{digits}{integer indicating the number of decimal places
    (\code{round}) or significant digits (\code{signif}) to be used.
    Negative values are allowed (see \sQuote{Details}).}

\item{...}{other arguments passed on to \code{\link{format}}}
}
\value{
A function that returns a character vector.
}
\description{
Comma and Significance Formatter
}
\examples{
comma_signif(1199)
}
\seealso{
\code{\link[scales]{comma}}
}
