% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dist_data}
\alias{dist_data}
\title{Distribution Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10 rows and 4 columns.
}
\usage{
dist_data
}
\description{
A data frame of information on the implemented distributions.
}
\details{
\describe{
\item{dist}{The distribution (chr).}
\item{npars}{The number of parameters (int).}
\item{tails}{Whether the distribution has both tails (flag).}
\item{stable}{Whether the distribution is numerically stable (flag).}
\item{bcanz}{Whether the distribution belongs to the set of distributions approved by BC, Canada, Australia and New Zealand for official guidelines (flag).}
}
}
\examples{
dist
}
\seealso{
Other dists: 
\code{\link{ssd_dists}()},
\code{\link{ssd_dists_all}()}
}
\concept{dists}
\keyword{datasets}
