% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{estimates.fitdists}
\alias{estimates.fitdists}
\title{Estimates for fitdists Object}
\usage{
\method{estimates}{fitdists}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused.}
}
\value{
A named list of the estimates.
}
\description{
Gets a named list of the estimated values by distribution and term.
}
\examples{
fits <- ssd_fit_dists(ssddata::ccme_boron)
estimates <- estimates(fits)
print(estimates)
ssd_hc(estimates)
ssd_plot_cdf(estimates)
}
\seealso{
\code{\link[=tidy.fitdists]{tidy.fitdists()}}, \code{\link[=ssd_match_moments]{ssd_match_moments()}}, \code{\link[=ssd_hc]{ssd_hc()}} and \code{\link[=ssd_plot_cdf]{ssd_plot_cdf()}}
}
