% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fluazinam_dists}
\alias{fluazinam_dists}
\title{fitdists for fitdistrplus fluazinam Data}
\format{
An object of class \code{fitdistscens} (inherits from \code{fitdists}) of length 3.
}
\usage{
fluazinam_dists
}
\description{
A fitdists object for Species Sensitivity Data for
Fluazinam.
}
\examples{
fluazinam_dists
}
\seealso{
\code{\link[fitdistrplus:fluazinam]{fitdistrplus::fluazinam()}}
}
\keyword{datasets}
