% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fluazinam_pred}
\alias{fluazinam_pred}
\title{Model averaged predictions for fluazinam}
\format{A tbl data frame:
\describe{
  \item{prop}{The proportion of species affected <dbl>.}
  \item{est}{The estimated concentration <dbl>.}
  \item{se}{The standard error of the estimate <dbl>.}
  \item{lcl}{The lower confidence limit <dbl>.}
  \item{se}{The upper confidence limit <dbl>.}
}}
\usage{
fluazinam_pred
}
\description{
A data frame of the predictions.
}
\examples{
head(fluazinam_pred)
}
\keyword{datasets}
