% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma.R
\name{gamma}
\alias{gamma}
\alias{sgamma}
\title{Gamma Distribution}
\usage{
sgamma(x)
}
\arguments{
\item{x}{A numeric vector of values.}
}
\value{
A numeric vector.
}
\description{
Starting values for the
Gamma distribution.
}
\examples{
x <- seq(0.01, 5, by = 0.01)
plot(x, dgamma(x, 1, 1), type = "l")
}
\seealso{
\code{\link[stats]{dgamma}}
}
