% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompertz.R
\name{gompertz}
\alias{gompertz}
\alias{dgompertz}
\alias{pgompertz}
\alias{qgompertz}
\alias{rgompertz}
\alias{sgompertz}
\title{gompertz Distribution}
\usage{
dgompertz(x, lshape = 0, lscale = 0, log = FALSE)

pgompertz(q, lshape = 0, lscale = 0, lower.tail = TRUE, log.p = FALSE)

qgompertz(p, lshape = 0, lscale = 0, lower.tail = TRUE, log.p = FALSE)

rgompertz(n, lshape = 0, lscale = 0)

sgompertz(x)
}
\arguments{
\item{x}{A numeric vector of values.}

\item{lshape}{shape parameter on the log scale.}

\item{lscale}{scale parameter on the log scale.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}

\item{q}{vector of quantiles.}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X <= x],otherwise, P[X > x].}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
A numeric vector.
}
\description{
Density, distribution function, quantile function and random generation for 
the gompertz distribution with parameters lshape and lscale.
}
\examples{
x <- seq(0.01, 5, by = 0.01)
plot(x, dgompertz(x), type = "l")
}
\seealso{
\code{\link[stats]{dgompertz}}
}
