% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.fitdists}
\alias{is.fitdists}
\title{Is fitdists Object}
\usage{
is.fitdists(x)
}
\arguments{
\item{x}{The object.}
}
\value{
A flag specifiying whether x is a fitdists Object.
}
\description{
Tests whether x is a fitdists Object.
}
\examples{
fits <- ssd_fit_dists(ssddata::ccme_boron)
is.fitdists(fits)
}
