% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.fitdists}
\alias{is.fitdists}
\title{Is fitdists}
\usage{
is.fitdists(x)
}
\arguments{
\item{x}{The object.}
}
\value{
A flag.
}
\description{
Tests whether an object is a fitdists.
}
\examples{
is.fitdists(boron_lnorm)
is.fitdists(boron_dists)
}
