% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censored.R
\name{is_censored}
\alias{is_censored}
\title{Is Censored
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
is_censored(x)
}
\arguments{
\item{x}{A fitdists object.}
}
\value{
A flag indicating if the data is censored.
}
\description{
Deprecated for \code{\link[=ssd_is_censored]{ssd_is_censored()}}.
}
\seealso{
\code{\link[=ssd_is_censored]{ssd_is_censored()}}
}
