% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgumbel.R
\name{lgumbel}
\alias{lgumbel}
\alias{dlgumbel}
\alias{qlgumbel}
\alias{plgumbel}
\alias{rlgumbel}
\alias{slgumbel}
\title{Log-Gumbel Distribution}
\usage{
dlgumbel(x, location = 0, scale = 0, log = FALSE)

qlgumbel(p, location = 0, scale = 0, lower.tail = TRUE, log.p = FALSE)

plgumbel(q, location = 0, scale = 0, lower.tail = TRUE, log.p = FALSE)

rlgumbel(n, location = 0, scale = 0)

slgumbel(x)
}
\arguments{
\item{x}{A numeric vector of values.}

\item{location}{location parameter.}

\item{scale}{scale parameter.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}

\item{p}{vector of probabilities.}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X <= x],otherwise, P[X > x].}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{q}{vector of quantiles.}

\item{n}{number of observations.}
}
\value{
A numeric vector.
}
\description{
Density, distribution function, quantile function, random generation
and starting values for the
Log-Gumbel distribution
with \code{scale} and \code{location} parameters.
}
