% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgumbel.R
\name{lgumbel}
\alias{lgumbel}
\alias{dlgumbel}
\alias{plgumbel}
\alias{qlgumbel}
\alias{rlgumbel}
\alias{slgumbel}
\title{Log-Gumbel Distribution}
\usage{
dlgumbel(x, llocation = 0, lscale = 1, log = FALSE)

plgumbel(q, llocation = 0, lscale = 1, lower.tail = TRUE, log.p = FALSE)

qlgumbel(p, llocation = 0, lscale = 1, lower.tail = TRUE, log.p = FALSE)

rlgumbel(n, llocation = 0, lscale = 1)

slgumbel(x)
}
\arguments{
\item{x}{A numeric vector of values.}

\item{llocation}{location parameter on the log scale.}

\item{lscale}{scale parameter on the log scale.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}

\item{q}{vector of quantiles.}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X <= x],otherwise, P[X > x].}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
A numeric vector.
}
\description{
Density, distribution function, quantile function, random generation
and starting values for the
Log-Gumbel distribution
with \code{lscale} and \code{llocation} parameters.
}
