% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnormlnorm.R
\name{lnormlnorm}
\alias{lnormlnorm}
\alias{dlnormlnorm}
\alias{plnormlnorm}
\alias{qlnormlnorm}
\alias{rlnormlnorm}
\title{Log Normal-Log Normal Mixture Distribution}
\usage{
dlnormlnorm(
  x,
  meanlog1 = 0,
  sdlog1 = 1,
  meanlog2 = 3,
  sdlog2 = 1,
  r = 0.5,
  log = FALSE
)

plnormlnorm(
  q,
  meanlog1 = 0,
  sdlog1 = 1,
  meanlog2 = 3,
  sdlog2 = 1,
  r = 0.5,
  lower.tail = TRUE,
  log.p = FALSE
)

qlnormlnorm(
  p,
  meanlog1 = 0,
  sdlog1 = 1,
  meanlog2 = 3,
  sdlog2 = 1,
  r = 0.5,
  lower.tail = TRUE,
  log.p = FALSE
)

rlnormlnorm(n, meanlog1 = 0, sdlog1 = 1, meanlog2 = 3, sdlog2 = 1, r = 0.5)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{meanlog1, sdlog1}{mean and standard deviation of the distribution on the log scale of the first mixture with default values of 0 and 1 respectively.}

\item{meanlog2, sdlog2}{mean and standard deviation of the distribution on the log scale of the second mixture with default values of 1 and 1 respectively.}

\item{r}{proportion of the first mixture.}

\item{log, log.p}{logical; if TRUE, probabilitise p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are \code{P[X <= x]}, otherwise, \code{P[X > x]}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. 
If length(n) > 1, the length is
taken to be the number required.}
}
\description{
Density, distribution function, quantile function and random generation
for the log-normal log-normal mixture distribution.
}
\examples{
hist(log(rlnormlnorm(1e+05)))
}
