% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nobs.fitdist}
\alias{nobs.fitdist}
\title{Number of Observations}
\usage{
\method{nobs}{fitdist}(object, ...)
}
\arguments{
\item{object}{The object.}

\item{...}{Unused.}
}
\description{
Number of Observations
}
\examples{
stats::nobs(boron_lnorm)
}
