% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{npars}
\alias{npars}
\title{Get the Number of Parameters}
\usage{
npars(object, ...)
}
\arguments{
\item{object}{The object.}

\item{...}{Unused.}
}
\value{
A count indicating the number of parameters.
}
\description{
Get the Number of Parameters
}
