% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npars.R
\name{npars}
\alias{npars}
\alias{npars.fitdist}
\alias{npars.fitdistcens}
\alias{npars.fitdists}
\title{Number of Parameters}
\usage{
npars(x, ...)

\method{npars}{fitdist}(x, ...)

\method{npars}{fitdistcens}(x, ...)

\method{npars}{fitdists}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused.}
}
\value{
A count indicating the number of parameters.
}
\description{
Get the Number of Parameters
}
\section{Methods (by class)}{
\itemize{
\item \code{fitdist}: Get the Number of parameters

\item \code{fitdistcens}: Get the Number of parameters

\item \code{fitdists}: Get the Number of parameters
}}

\examples{
npars(boron_lnorm)
npars(boron_dists)
npars(fluazinam_lnorm)
npars(fluazinam_dists)
}
