% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameter Descriptions for ssdtools Functions}
\arguments{
\item{x}{The object.}

\item{conc}{A numeric vector of the concentrations.}

\item{ic}{A string specifying the information-theoretic criterion to use for model averaging ('aic', 'aicc' or 'bic').}

\item{ci}{A flag specifying whether to estimate confidence intervals by parametric bootstrapping.}

\item{...}{Unused.}
}
\description{
Parameter Descriptions for ssdtools Functions
}
\keyword{internal}
