% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameter Descriptions for ssdtools Functions}
\arguments{
\item{x}{The object.}

\item{conc}{A numeric vector of the concentrations.}

\item{ic}{A string specifying the information-theoretic criterion to use for model averaging ('aic', 'aicc' or 'bic').}

\item{ci}{A flag specifying whether to estimate confidence intervals by parametric bootstrapping.}

\item{...}{Unused.}

\item{q}{vector of quantiles.}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X <= x],otherwise, P[X > x].}

\item{location}{location parameter.}

\item{locationlog}{location parameter on the log scale.}

\item{shape}{shape parameter.}

\item{scale}{scale parameter.}

\item{scalelog}{scale parameter on the log scale.}

\item{shapelog}{shape parameter on the log scale.}

\item{shape1log}{shape1 parameter on the log scale.}

\item{shape2log}{shape2 parameter on the log scale.}
}
\description{
Parameter Descriptions for ssdtools Functions
}
\keyword{internal}
