% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.fitdist}
\alias{predict.fitdist}
\title{Predict fitdist}
\usage{
\method{predict}{fitdist}(object, probs = seq(0.01, 0.99, by = 0.02),
  nboot = 1001, level = 0.95, ...)
}
\arguments{
\item{object}{The object.}

\item{probs}{A numeric vector of the densities to calculate the estimated concentrations for.}

\item{nboot}{A count of the number of bootstrap samples to use to estimate the se and confidence limits.}

\item{level}{A real of the confidence limit.}

\item{...}{Unused}
}
\description{
Predict fitdist
}
\examples{
predict(boron_lnorm)
}
