% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.fitdist}
\alias{predict.fitdist}
\title{Predict fitdist}
\usage{
\method{predict}{fitdist}(
  object,
  percent = 1:99,
  ci = FALSE,
  level = 0.95,
  nboot = 1000,
  parallel = NULL,
  ncpus = 1,
  na.rm = FALSE,
  ...
)
}
\arguments{
\item{object}{The object.}

\item{percent}{A numeric vector of percentages.}

\item{ci}{A flag specifying whether to estimate confidence intervals (by parametric bootstrapping).}

\item{level}{A number between 0 and 1 of the confidence level.}

\item{nboot}{A count of the number of bootstrap samples to use to estimate the se and confidence limits.}

\item{parallel}{A string specifying the type of parallel operation to be used ('no', 'snow' or 'multicore').}

\item{ncpus}{A count of the number of parallel processes to use.}

\item{na.rm}{A flag specifying whether to silently remove missing values or
remove them with a warning.}

\item{...}{Unused.}
}
\description{
Predict fitdist
}
\examples{
predict(boron_lnorm, percent = c(5L, 50L))
}
