% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.fitdistcens}
\alias{predict.fitdistcens}
\title{Predict censored fitdist}
\usage{
\method{predict}{fitdistcens}(object, props = seq(0.01, 0.99, by = 0.01),
  nboot = 1001, level = 0.95, ...)
}
\arguments{
\item{object}{The object.}

\item{props}{A numeric vector of the densities to calculate the estimated concentrations for.}

\item{nboot}{A count of the number of bootstrap samples to use to estimate the se and confidence limits.}

\item{level}{The confidence level.}

\item{...}{Unused}
}
\description{
Predict censored fitdist
}
\examples{
predict(fluazinam_lnorm, props = c(0.05, 0.5))
}
