% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.fitdistcens}
\alias{predict.fitdistcens}
\title{Predict censored fitdist}
\usage{
\method{predict}{fitdistcens}(object, percent = 1:99, nboot = 1000,
  level = 0.95, ...)
}
\arguments{
\item{object}{The object.}

\item{percent}{A numeric vector of the densities to calculate the estimated concentrations for.}

\item{nboot}{A count of the number of bootstrap samples to use to estimate the se and confidence limits.}

\item{level}{The confidence level.}

\item{...}{Unused}
}
\description{
Predict censored fitdist
}
\examples{
\dontrun{
predict(fluazinam_lnorm, percent = c(5L, 50L))
}
}
