% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.fitdists}
\alias{predict.fitdists}
\title{Predict fitdists}
\usage{
\method{predict}{fitdists}(
  object,
  percent = 1:99,
  nboot = 1000,
  ic = "aicc",
  average = TRUE,
  level = 0.95,
  parallel = "no",
  ncpus = 1L,
  ...
)
}
\arguments{
\item{object}{The object.}

\item{percent}{A numeric vector of the densities to calculate the estimated concentrations for.}

\item{nboot}{A count of the number of bootstrap samples to use to estimate the se and confidence limits.}

\item{ic}{A string indicating which information-theoretic criterion ('aic', 'aicc' or 'bic') to use for model averaging.}

\item{average}{A flag indicating whether to model-average.}

\item{level}{The confidence level.}

\item{parallel}{The type of parallel operation to be used, \code{"snow"} or \code{"multicore"}
  (the second one not being available on Windows),
  or \code{"no"} if no parallel operation.}

\item{ncpus}{Number of processes to be used in parallel operation : 
  typically one would fix it to the number of available CPUs.}

\item{...}{Unused}
}
\description{
Predict fitdists
}
\examples{
\dontrun{
predict(boron_dists)
}
}
