% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.fitdists}
\alias{predict.fitdists}
\title{Predict fitdist}
\usage{
\method{predict}{fitdists}(object, probs = seq(0.01, 0.99, by = 0.01),
  nboot = 1001, ic = "aicc", average = TRUE, level = 0.95, ...)
}
\arguments{
\item{object}{The object.}

\item{probs}{A numeric vector of the densities to calculate the estimated concentrations for.}

\item{nboot}{A count of the number of bootstrap samples to use to estimate the se and confidence limits.}

\item{ic}{A string indicating which information-theoretic criterion ('aic', 'aicc' or 'bic') to use for model averaging or use 'no' to not model average.}

\item{average}{A flag indicating whether to model-average.}

\item{level}{A real of the confidence limit.}

\item{...}{Unused}
}
\description{
Predict fitdist
}
\examples{
\dontrun{
predict(boron_dists)
}
}
