% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.fitdists}
\alias{predict.fitdists}
\title{Predict fitdist}
\usage{
\method{predict}{fitdists}(object, probs = seq(0.01, 0.99, by = 0.02),
  nboot = 1001, IC = AICc, level = 0.95, ...)
}
\arguments{
\item{object}{The object.}

\item{probs}{A numeric vector of the densities to calculate the estimated concentrations for.}

\item{nboot}{A count of the number of bootstrap samples to use to estimate the se and confidence limits.}

\item{IC}{A function returning the information criterion to use for model averaging.}

\item{level}{A real of the confidence limit.}

\item{...}{Unused}
}
\description{
Predict fitdist
}
\examples{
\dontrun{
predict(boron_dists, probs = c(0.01, 0.05, 0.5), IC = AICc)
}
}
\seealso{
\code{\link{AICc}}
}
