% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.fitdists}
\alias{predict.fitdists}
\title{Predict fitdist}
\usage{
\method{predict}{fitdists}(
  object,
  percent = 1:99,
  nboot = 1000,
  ic = "aicc",
  average = TRUE,
  level = 0.95,
  ...
)
}
\arguments{
\item{object}{The object.}

\item{percent}{A numeric vector of the densities to calculate the estimated concentrations for.}

\item{nboot}{A count of the number of bootstrap samples to use to estimate the se and confidence limits.}

\item{ic}{A string indicating which information-theoretic criterion ('aic', 'aicc' or 'bic') to use for model averaging.}

\item{average}{A flag indicating whether to model-average.}

\item{level}{The confidence level.}

\item{...}{Unused}
}
\description{
Predict fitdist
}
\examples{
\dontrun{
predict(boron_dists)
}
}
