% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.fitdistscens}
\alias{predict.fitdistscens}
\title{Predict censored fitdists}
\usage{
\method{predict}{fitdistscens}(
  object,
  percent = 1:99,
  ci = FALSE,
  level = 0.95,
  nboot = 1000,
  parallel = NULL,
  ncpus = 1,
  na.rm = FALSE,
  average = TRUE,
  ic = "aic",
  ...
)
}
\arguments{
\item{object}{The object.}

\item{percent}{A numeric vector of percentages.}

\item{ci}{A flag specifying whether to estimate confidence intervals (by parametric bootstrapping).}

\item{level}{A number between 0 and 1 of the confidence level.}

\item{nboot}{A count of the number of bootstrap samples to use to estimate the se and confidence limits.}

\item{parallel}{A string specifying the type of parallel operation to be used ('no', 'snow' or 'multicore').}

\item{ncpus}{A count of the number of parallel processes to use.}

\item{na.rm}{A flag specifying whether to silently remove missing values or
remove them with a warning.}

\item{average}{A flag specifying whether to model average the estimates.}

\item{ic}{A string specifying which information-theoretic criterion ('aic', 'aicc' or 'bic') to use for model averaging .}

\item{...}{Unused.}
}
\description{
Predict censored fitdists
}
\examples{
predict(fluazinam_dists)
}
