% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.fitdistscens}
\alias{predict.fitdistscens}
\title{Predict Censored fitdists}
\usage{
\method{predict}{fitdistscens}(object, props = seq(0.01, 0.99, by = 0.01),
  nboot = 1001, ic = "aic", average = TRUE, level = 0.95, ...)
}
\arguments{
\item{object}{The object.}

\item{props}{A numeric vector of the densities to calculate the estimated concentrations for.}

\item{nboot}{A count of the number of bootstrap samples to use to estimate the se and confidence limits.}

\item{ic}{A string indicating which information-theoretic criterion ('aic', 'aicc' or 'bic') to use for model averaging.}

\item{average}{A flag indicating whether to model-average.}

\item{level}{The confidence level.}

\item{...}{Unused}
}
\description{
Predict Censored fitdists
}
\examples{
\dontrun{
predict(fluazinam_dists)
}
}
