% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R, R/coef.R, R/estimates.R,
%   R/glance.R, R/logLik.R, R/nobs.R, R/npars.R, R/plot.R, R/predict.R, R/tidy.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{autoplot}
\alias{coef}
\alias{estimates}
\alias{glance}
\alias{logLik}
\alias{nobs}
\alias{npars}
\alias{plot}
\alias{predict}
\alias{tidy}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{glance}}, \code{\link[generics]{tidy}}}

  \item{ggplot2}{\code{\link[ggplot2]{autoplot}}}

  \item{graphics}{\code{\link[graphics:plot.default]{plot}}}

  \item{stats}{\code{\link[stats]{coef}}, \code{\link[stats]{logLik}}, \code{\link[stats]{nobs}}, \code{\link[stats]{predict}}}

  \item{universals}{\code{\link[universals]{estimates}}, \code{\link[universals]{npars}}}
}}

