% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R, R/glance.R, R/logLik.R, R/nobs.R,
%   R/npars.R, R/tidy.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{augment}
\alias{glance}
\alias{logLik}
\alias{nobs}
\alias{npars}
\alias{tidy}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{augment}}, \code{\link[generics]{glance}}, \code{\link[generics]{tidy}}}

  \item{stats}{\code{\link[stats]{logLik}}, \code{\link[stats]{nobs}}}

  \item{universals}{\code{\link[universals]{npars}}}
}}

