% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{scale_colour_ssd}
\alias{scale_colour_ssd}
\alias{scale_color_ssd}
\title{Discrete color-blind scale for SSD Plots}
\usage{
scale_colour_ssd(...)

scale_color_ssd(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.}
}
\description{
Discrete color-blind scale for SSD Plots
}
\section{Functions}{
\itemize{
\item \code{scale_color_ssd}: Discrete color-blind scale for SSD Plots
}}

\examples{
ssd_plot(boron_data, boron_pred, shape = "Group") +
  scale_colour_ssd()
}
\seealso{
Other ggplot: 
\code{\link{geom_hcintersect}()},
\code{\link{geom_ssdpoint}()},
\code{\link{geom_ssdsegment}()},
\code{\link{geom_xribbon}()},
\code{\link{ssd_pal}()}
}
\concept{ggplot}
