% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssd-data.R
\name{ssd_data}
\alias{ssd_data}
\title{Species Sensitivity Data}
\usage{
ssd_data(x)
}
\arguments{
\item{x}{A fitdists object.}
}
\value{
A tbl data frame of the data.
}
\description{
Species Sensitivity Data
}
\examples{
fits <- ssd_fit_dists(ssdtools::boron_data, dists = "lnorm")
ssd_data(fits)
}
