% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcanz.R
\name{ssd_dists_bcanz}
\alias{ssd_dists_bcanz}
\title{BCANZ Distributions}
\usage{
ssd_dists_bcanz()
}
\value{
A unique, sorted character vector of the distributions.
}
\description{
Gets a character vector of the names of the distributions
adopted by BC, Canada, Australia and New Zealand for official guidelines.
}
\examples{
ssd_dists_bcanz()
}
\seealso{
\code{\link[=ssd_dists]{ssd_dists()}}
}
\concept{dists BCANZ}
