% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ssd_ecd}
\alias{ssd_ecd}
\title{Empirical Cumulative Density}
\usage{
ssd_ecd(x, ties.method = "first")
}
\arguments{
\item{x}{a numeric, complex, character or logical vector.}

\item{ties.method}{a character string specifying how ties are treated,
    see \sQuote{Details}; can be abbreviated.}
}
\value{
A numeric vector of the empirical cumulative density.
}
\description{
Empirical Cumulative Density
}
\examples{
ssd_ecd(1:10)
}
