% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ssd_ecd}
\alias{ssd_ecd}
\title{Empirical Cumulative Density}
\usage{
ssd_ecd(x)
}
\arguments{
\item{x}{A numeric vector of species sensitivity values.}
}
\value{
A numeric vector of the empirical cumulative density.
}
\description{
Empirical Cumulative Density
}
\examples{
ssd_ecd(1:10)
}
