% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ssd_ecdf}
\alias{ssd_ecdf}
\title{Empirical Cumulative Distribution Function}
\usage{
ssd_ecdf(x, ties.method = "first")
}
\arguments{
\item{x}{a numeric, complex, character or logical vector.}

\item{ties.method}{a character string specifying how ties are treated,
    see \sQuote{Details}; can be abbreviated.}
}
\value{
A numeric vector of the empirical cumulative distribution function.
}
\description{
Empirical Cumulative Distribution Function
}
\examples{
ssd_ecdf(1:10)
}
