% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{ssd_fit_dists}
\alias{ssd_fit_dists}
\title{Fit Distributions}
\usage{
ssd_fit_dists(data, left = "Conc", right = left, weights = NULL,
  dists = c("lnorm", "llog", "gompertz", "lgumbel", "gamma", "weibull"),
  silent = FALSE)
}
\arguments{
\item{data}{A data frame.}

\item{left}{A string of the column in data with the left concentration values.}

\item{right}{A string of the column in data with the right concentration values.}

\item{weights}{A string of the column in data with the weightings (or NULL)}

\item{dists}{A character vector of the distributions to fit.}

\item{silent}{A flag indicating whether fits should fail silently.}
}
\value{
An object of class fitdists (a list of \code{\link[fitdistrplus]{fitdist}} objects).
}
\description{
Fits one or more distributions to species sensitivity data.
}
\details{
By default the 'lnorm', 'llog', 'gompertz', 'lgumbel', 'gamma' and 'weibull'
distributions are fitted to the data.
The ssd_fit_dist and \code{\link{ssd_fit_dist}} functions have also been
tested with the 'burr' and 'pareto' distributions.
}
\examples{
ssd_fit_dists(boron_data)
}
\seealso{
\code{\link{ssd_fit_dist}}
}
