% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{ssd_fit_dists}
\alias{ssd_fit_dists}
\title{Fit Distributions}
\usage{
ssd_fit_dists(
  data,
  left = "Conc",
  right = left,
  weight = NULL,
  dists = c("gamma", "llogis", "lnorm"),
  nrow = 6L,
  rescale = FALSE,
  computable = TRUE,
  silent = FALSE
)
}
\arguments{
\item{data}{A data frame.}

\item{left}{A string of the column in data with the concentrations.}

\item{right}{A string of the column in data with the right concentration values.}

\item{weight}{A string of the numeric column in data with positive weights less than or equal to 1,000 or NULL.}

\item{dists}{A character vector of the distribution names.}

\item{nrow}{A positive whole number of the minimum number of non-missing rows.}

\item{rescale}{A flag specifying whether to rescale concentration values by dividing by the largest finite value.}

\item{computable}{A flag specifying whether to only return fits with numerically computable standard errors.}

\item{silent}{A flag indicating whether fits should fail silently.}
}
\value{
An object of class fitdists.
}
\description{
Fits one or more distributions to species sensitivity data.
}
\details{
By default the 'llogis', 'gamma' and 'lnorm'
distributions are fitted to the data.
The ssd_fit_dists function has also been
tested with the 'gompertz', 'lgumbel' and 'weibull' distributions.

If weight specifies a column in the data frame with positive numbers,
weighted estimation occurs.
However, currently only the resultant parameter estimates are available (via coef).

If the \code{right} argument is different to the \code{left} argument then the data are considered to be censored.
}
\examples{
ssd_fit_dists(boron_data)
}
