% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-dists.R
\name{ssd_fit_dists}
\alias{ssd_fit_dists}
\title{Fit Distributions}
\usage{
ssd_fit_dists(x, dists = c("lnorm", "llog", "gompertz", "lgumbel", "gamma",
  "weibull"), silent = FALSE)
}
\arguments{
\item{x}{A numeric vector of the data.}

\item{dists}{A character vector of the distributions to fit.}

\item{silent}{A flag indicating whether fits should fail without issuing errors.}
}
\value{
An object of class fitdists.
}
\description{
By default fits xx distributions.
}
\examples{
ssd_fit_dists(ccme_data$Conc[ccme_data$Chemical == "Boron"])
}
