% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{ssd_gof}
\alias{ssd_gof}
\title{Goodness of Fit}
\usage{
ssd_gof(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused.}
}
\value{
A tbl data frame of the gof statistics.
}
\description{
Returns a tbl data frame with the following columns
\describe{
  \item{dist}{The distribution name (chr)}
  \item{ad}{Anderson-Darling statistic (dbl)}
  \item{ks}{Kolmogorov-Smirnov statistic (dbl)}
  \item{cvm}{Cramer-von Mises statistic (dbl)}
  \item{aic}{Akaike's Information Criterion (dbl)}
  \item{aicc}{Akaike's Information Criterion corrected for sample size (dbl)}
  \item{bic}{Bayesion Information Criterion (dbl)}
}
}
\examples{
ssd_gof(boron_lnorm)
ssd_gof(boron_dists)
}
