% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{ssd_gof}
\alias{ssd_gof}
\title{Goodness of Fit}
\usage{
ssd_gof(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused.}
}
\value{
A tbl data frame of the gof statistics.
}
\description{
Goodness of Fit
}
\examples{
ssd_gof(boron_lnorm)
ssd_gof(boron_dists)
}
