% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc5.R
\name{ssd_hc5}
\alias{ssd_hc5}
\title{Get HC5}
\usage{
ssd_hc5(object, ...)
}
\arguments{
\item{object}{The object.}

\item{...}{Unused.}
}
\value{
A data frame of the hc5 estimate with se and upper and lower confidence intervals.
}
\description{
Get HC5
}
\examples{
ssd_hc5(boron_lnorm)
}
