% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc-burrlioz.R
\name{ssd_hc_burrlioz}
\alias{ssd_hc_burrlioz}
\title{Hazard Concentrations for Burrlioz Fit
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
ssd_hc_burrlioz(
  x,
  percent = 5,
  ci = FALSE,
  level = 0.95,
  nboot = 1000,
  min_pboot = 0.99,
  parametric = FALSE
)
}
\arguments{
\item{x}{The object.}

\item{percent}{A numeric vector of percent values to estimate hazard concentrations for.}

\item{ci}{A flag specifying whether to estimate confidence intervals (by bootstrapping).}

\item{level}{A number between 0 and 1 of the confidence level.}

\item{nboot}{A count of the number of bootstrap samples to use to estimate confidence limits. A value of at least 10,000 is recommended for official guidelines.}

\item{min_pboot}{A number between 0 and 1 of the minimum
proportion of bootstrap samples that must successfully fit in the sense of returning a likelihood
in order to report the confidence intervals.}

\item{parametric}{A flag specifying whether to perform parametric (as opposed to non-parametric) bootstrapping when \code{ci = TRUE}.}
}
\value{
A tibble of corresponding hazard concentrations.
}
\description{
Deprecated for \code{\link[=ssd_hc]{ssd_hc()}}.
}
\examples{
fit <- ssd_fit_burrlioz(ssddata::ccme_boron)
ssd_hc_burrlioz(fit)

}
