% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hp.R
\name{ssd_hp}
\alias{ssd_hp}
\alias{ssd_hp.fitdist}
\alias{ssd_hp.fitdists}
\alias{ssd_hp.fitdistcens}
\alias{ssd_hp.fitdistscens}
\title{Percent Protected}
\usage{
ssd_hp(x, ...)

\method{ssd_hp}{fitdist}(x, conc, ...)

\method{ssd_hp}{fitdists}(x, conc, ic = "aicc", ...)

\method{ssd_hp}{fitdistcens}(x, conc, ...)

\method{ssd_hp}{fitdistscens}(x, conc, ic = "aic", ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused.}

\item{conc}{A numeric vector of the concentrations.}

\item{ic}{A string indicating which information-theoretic criterion ('aic', 'aicc' or 'bic') to use for model averaging.}
}
\value{
A data frame of the conc and percent.
}
\description{
Gets percent species protected at specified concentrations.
}
\section{Methods (by class)}{
\itemize{
\item \code{fitdist}: Percent Protected fitdist

\item \code{fitdists}: Percent Protected fitdists

\item \code{fitdistcens}: Percent Protected fitdistcens

\item \code{fitdistscens}: Percent Protected fitdistcens
}}

\examples{
ssd_hp(boron_lnorm, c(0, 1, 30, Inf))
ssd_hp(boron_dists, c(0, 1, 30, Inf))
ssd_hp(fluazinam_lnorm, c(0, 1, 30, Inf))
ssd_hp(fluazinam_dists, c(0, 1, 30, Inf))
}
