% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hp.R
\name{ssd_hp}
\alias{ssd_hp}
\alias{ssd_hp.fitdists}
\title{Hazard Percent}
\usage{
ssd_hp(x, ...)

\method{ssd_hp}{fitdists}(
  x,
  conc,
  ci = FALSE,
  level = 0.95,
  nboot = 1000,
  control = NULL,
  parallel = NULL,
  ncpus = 1,
  average = TRUE,
  ic = "aicc",
  ...
)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused.}

\item{conc}{A numeric vector of concentrations.}

\item{ci}{A flag specifying whether to estimate confidence intervals (by parametric bootstrapping).}

\item{level}{A number between 0 and 1 of the confidence level.}

\item{nboot}{A count of the number of bootstrap samples to use to estimate the se and confidence limits.}

\item{control}{A list of control parameters passed to \link[stats:optim]{stats::optim}.}

\item{parallel}{A string specifying the type of parallel operation to be used ('no', 'snow' or 'multicore').}

\item{ncpus}{A count of the number of parallel processes to use.}

\item{average}{A flag specifying whether to model average the estimates.}

\item{ic}{A string specifying which information-theoretic criterion ('aic', 'aicc' or 'bic') to use for model averaging .}
}
\value{
A tibble of corresponding hazard percents.
}
\description{
Gets percent of species protected at specified concentration(s).
}
\details{
If \code{ci = TRUE} uses parameteric bootstrapping to get confidence intervals on the
hazard percent(s).
}
\section{Methods (by class)}{
\itemize{
\item \code{fitdists}: Hazard Percents for fitdists Object
}}

\examples{
fits <- ssd_fit_dists(ssdtools::boron_data)
ssd_hp(fits, conc = 1)
}
\seealso{
\code{\link[=ssd_hc]{ssd_hc()}}
}
