% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hp.R
\name{ssd_hp}
\alias{ssd_hp}
\alias{ssd_hp.fitdist}
\alias{ssd_hp.fitdists}
\alias{ssd_hp.fitdistcens}
\alias{ssd_hp.fitdistscens}
\title{Hazard Percent}
\usage{
ssd_hp(x, ...)

\method{ssd_hp}{fitdist}(x, conc, ci = FALSE, ...)

\method{ssd_hp}{fitdists}(x, conc, ic = "aicc", ci = FALSE, ...)

\method{ssd_hp}{fitdistcens}(x, conc, ci = FALSE, ...)

\method{ssd_hp}{fitdistscens}(x, conc, ic = "aic", ci = FALSE, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused.}

\item{conc}{A numeric vector of the concentrations.}

\item{ci}{A flag specifying whether to estimate confidence intervals by parametric bootstrapping.}

\item{ic}{A string specifying the information-theoretic criterion to use for model averaging ('aic', 'aicc' or 'bic').}
}
\value{
A data frame of the conc and percent.
}
\description{
Gets percent species protected at specified concentrations.
}
\section{Methods (by class)}{
\itemize{
\item \code{fitdist}: Hazard Percent fitdist

\item \code{fitdists}: Hazard Percent fitdists

\item \code{fitdistcens}: Hazard Percent fitdistcens

\item \code{fitdistscens}: Hazard Percent fitdistcens
}}

\examples{
ssd_hp(boron_lnorm, c(0, 1, 30, Inf))
ssd_hp(boron_dists, c(0, 1, 30, Inf))
ssd_hp(fluazinam_lnorm, c(0, 1, 30, Inf))
ssd_hp(fluazinam_dists, c(0, 1, 30, Inf))
}
