% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match-moments.R
\name{ssd_match_moments}
\alias{ssd_match_moments}
\title{Match Moments}
\usage{
ssd_match_moments(
  dists = ssd_dists("bc"),
  meanlog = 1,
  sdlog = 1,
  nsim = 1e+05
)
}
\arguments{
\item{dists}{A character vector of the distribution names.}

\item{meanlog}{A number of the mean on the log scale.}

\item{sdlog}{A number of the standard deviation on the log scale.}

\item{nsim}{A positive whole number of the number of simulations to generate.}
}
\value{
A named list of the parameter values that produce a distribution
with moments closest to the meanlog and sdlog.
}
\description{
Match Moments
}
\examples{
ssd_match_moments()
}
\seealso{
\code{\link[=ssd_plot_cdf]{ssd_plot_cdf()}}.
}
