% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{ssd_moments}
\alias{ssd_moments}
\title{Calculate Moments}
\usage{
ssd_moments(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A tbl data frame with 1 row and the columns min, max, median, mean, sd, skewness, kurtosis.
}
\description{
Calculates the min, max, median and unbiased estimates of the first four moments.
}
\examples{
ssd_moments(ccme_data$Conc[ccme_data$Chemical == "Cadmium"])
}
\seealso{
\code{\link[fitdistrplus]{descdist}}
}
