% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ssd_plot}
\alias{ssd_plot}
\title{SSD Plot}
\usage{
ssd_plot(data, pred, conc = "Concentration", label = NULL, shape = NULL,
  xlab = "Concentration", ylab = "Percent of Species Affected", ci = TRUE,
  hc5 = TRUE)
}
\arguments{
\item{data}{A data frame.}

\item{pred}{A data frame of the predictions.}

\item{conc}{A string of the column in data with the concentrations.}

\item{label}{A string of the column in data with the labels.}

\item{shape}{A string of the column in data for the shape aesthetic.}

\item{xlab}{A string of the x-axis label.}

\item{ylab}{A string of the x-axis label.}

\item{ci}{A flag indicating wether to plot confidence intervals}

\item{hc5}{A flag indicating whether to plot the HC5 (not yet implemented)}
}
\description{
SSD Plot
}
\examples{
ssd_plot(boron_data, boron_pred, label = "Species", shape = "Group")
}
