% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-cdf.R
\name{ssd_plot_cdf}
\alias{ssd_plot_cdf}
\alias{ssd_plot_cdf.fitdists}
\alias{ssd_plot_cdf.list}
\title{Plot Cumulative Distribution Function (CDF)}
\usage{
ssd_plot_cdf(x, ...)

\method{ssd_plot_cdf}{fitdists}(x, average = FALSE, ...)

\method{ssd_plot_cdf}{list}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused.}

\item{average}{A flag specifying whether to model average the estimates.}
}
\description{
Generic function to plots the cumulative distribution function (CDF).
}
\section{Methods (by class)}{
\itemize{
\item \code{fitdists}: Plot CDF for fitdists object

\item \code{list}: Plot CDF for named list of distributional parameter values
}}

\examples{
fits <- ssd_fit_dists(ssdtools::boron_data)
ssd_plot_cdf(fits)

ssd_plot_cdf(list(
  llogis = c(locationlog = 2, scalelog = 1),
  lnorm = c(meanlog = 2, sdlog = 2))
)
}
\seealso{
\code{\link[=ssd_plot]{ssd_plot()}}
}
