% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ssd_plot_cdf}
\alias{ssd_plot_cdf}
\title{Cumulative Distribution Function Plot}
\usage{
ssd_plot_cdf(data, xlab = "Concentration", ylab = "Species Affected")
}
\arguments{
\item{data}{A data frame.}

\item{xlab}{A string of the x-axis label.}

\item{ylab}{A string of the x-axis label.}
}
\description{
Plots the cdf for a dists object.
}
\examples{
ssd_plot_cdf(boron_dists)
}
