% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-cf.R
\name{ssd_plot_cf}
\alias{ssd_plot_cf}
\title{Cullen and Frey Plot
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
ssd_plot_cf(data, left = "Conc")
}
\arguments{
\item{data}{A data frame.}

\item{left}{A string of the column in data with the concentrations.}
}
\description{
Plots a Cullen and Frey graph of the skewness and kurtosis
for non-censored data.
}
\details{
Deprecated for \code{\link[fitdistrplus:descdist]{fitdistrplus::descdist()}}.
}
