% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-cf.R
\name{ssd_plot_cf}
\alias{ssd_plot_cf}
\alias{ssd_cfplot}
\title{Cullen and Frey Plot}
\usage{
ssd_plot_cf(data, left = "Conc")

ssd_cfplot(data, left = "Conc")
}
\arguments{
\item{data}{A data frame.}

\item{left}{A string of the column in data with the concentrations.}
}
\description{
Plots a Cullen and Frey graph of the skewness and kurtosis
for non-censored data.
}
\section{Functions}{
\itemize{
\item \code{ssd_cfplot}: Deprecated Cullen and Frey Plot
}}

\examples{
ssd_plot_cf(boron_data)
}
\seealso{
\code{\link[fitdistrplus:descdist]{fitdistrplus::descdist()}}
}
