% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp.R
\name{ssd_pp}
\alias{ssd_pp}
\alias{ssd_pp.fitdist}
\alias{ssd_pp.fitdists}
\alias{ssd_pp.fitdistcens}
\title{Percent Protected}
\usage{
ssd_pp(x, ...)

\method{ssd_pp}{fitdist}(x, conc, ...)

\method{ssd_pp}{fitdists}(x, conc, ic = "aicc", ...)

\method{ssd_pp}{fitdistcens}(x, conc, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused.}

\item{conc}{A numeric vector of the concentrations.}

\item{ic}{A string indicating which information-theoretic criterion ('aic', 'aicc' or 'bic') to use for model averaging.}
}
\value{
A data frame of the conc and percent.
}
\description{
Gets percent species protected at specified concentrations.
}
\section{Methods (by class)}{
\itemize{
\item \code{fitdist}: Percent Protected fitdist

\item \code{fitdists}: Percent Protected fitdists

\item \code{fitdistcens}: Percent Protected fitdistcens
}}

\examples{
ssd_pp(boron_lnorm, c(0, 1, 30, Inf))
ssd_pp(boron_dists, c(0, 1, 30, Inf))
ssd_pp(fluazinam_lnorm, c(0, 1, 30, Inf))
}
