% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggproto.R
\docType{data}
\name{ssdtools-ggproto}
\alias{ssdtools-ggproto}
\alias{StatSsd}
\alias{StatSsdcens}
\alias{GeomSsd}
\alias{GeomSsdcens}
\alias{GeomHcintersect}
\alias{GeomXribbon}
\title{Base ggproto classes for ggplot2}
\description{
Base ggproto classes for ggplot2
}
\seealso{
\code{\link[ggplot2:ggplot2-ggproto]{ggplot2::ggplot2-ggproto()}}
}
\keyword{datasets}
