% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssdtools-package.R
\docType{package}
\name{ssdtools-package}
\alias{ssdtools}
\alias{ssdtools-package}
\title{ssdtools: Species Sensitivity Distributions}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Species sensitivity distributions are cumulative probability 
  distributions which are fitted to toxicity concentrations for multiple species.
  The ssdtools package uses Maximum Likelihood to fit
  log-normal, log-logistic, log-gumbel, gompertz, gamma or weibull distributions.
  Multiple distributions can be averaged using Information Criteria.
  Confidence intervals on fits and hazard concentrations 
  are produced by bootstrapping.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bcgov/ssdca}
  \item Report bugs at \url{https://github.com/bcgov/ssdca/issues}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (0000-0002-7683-4592) [contractor]

Authors:
\itemize{
  \item Carl Schwarz [contractor]
}

Other contributors:
\itemize{
  \item Ali Azizishirazi \email{Ali.Azizishirazi@gov.bc.ca} [contributor]
  \item Angeline Tillmanns [contributor]
  \item Stephanie Hazlitt [contributor]
  \item Province of British Columbia [copyright holder]
}

}
\keyword{internal}
