% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssdtools-package.R
\docType{package}
\name{ssdtools-package}
\alias{ssdtools}
\alias{ssdtools-package}
\title{ssdtools: Species Sensitivity Distributions}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Species sensitivity distributions are 
  cumulative probability distributions which are fitted to 
  toxicity concentrations for different species.
  The ssdtools package uses Maximum Likelihood to fit distributions 
  such as the log-normal, gamma, burr Type-III, log-logistic, 
  log-Gumbel, Gompertz and Weibull.
  The user can provide custom distributions.
  Multiple distributions can be averaged using Information Criteria.
  Confidence intervals on hazard concentrations and proportions are produced by 
  parametric bootstrapping.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bcgov/ssdtools}
  \item Report bugs at \url{https://github.com/bcgov/ssdtools/issues}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID}) [contractor]

Authors:
\itemize{
  \item Carl Schwarz [contractor]
}

Other contributors:
\itemize{
  \item Ali Azizishirazi \email{Ali.Azizishirazi@gov.bc.ca} [contributor]
  \item Angeline Tillmanns [contributor]
  \item Stephanie Hazlitt [contributor]
  \item Kathleen McTavish [contributor]
  \item Emilie Doussantousse [contributor]
  \item Heather Thompson [contributor]
  \item Andy Teucher [contributor]
  \item Nan-Hung Hsieh [contributor]
  \item Sergio Ibarra Espinosa [contributor]
  \item Province of British Columbia [copyright holder]
}

}
\keyword{internal}
