% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset.fitdists}
\alias{subset.fitdists}
\title{Subset fitdists Object}
\usage{
\method{subset}{fitdists}(x, select = names(x), ...)
}
\arguments{
\item{x}{The object.}

\item{select}{A character vector of the distributions to select.}

\item{...}{Unused.}
}
\description{
Select a subset of distributions from a fitdists object.
}
\examples{
fits <- ssd_fit_dists(ssdtools::boron_data)
subset(fits, c("gamma", "lnorm"))
}
