% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset.fitdists}
\alias{subset.fitdists}
\title{Subset fitdists}
\usage{
\method{subset}{fitdists}(x, select = names(x), ...)
}
\arguments{
\item{x}{The object.}

\item{select}{A character vector of the distributions to select.}

\item{...}{Unused.}
}
\description{
Subset fitdists
}
\examples{
subset(boron_dists, c("gamma", "lnorm"))
}
