% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{Test Data}
\format{
A tbl data frame:
\describe{
  \item{Chemical}{The chemical <chr>.}
  \item{Conc}{The chemical concentration <dbl>.}
}
}
\usage{
test_data
}
\description{
Data to test ssdtools.
}
\examples{
head(test_data)
}
\keyword{datasets}
